@@ -147,58 +162,44 @@ class sdist(setuptools.command.sdist.sdist):
 
 
 setuptools.setup(
-    name = project_name,
-    version = project_version,
-    description = "An extensible microscope hardware interface.",
-    long_description = open('README', 'r').read(),
-    license = "GPL-3.0+",
-
-    ## We need an author and an author_email value or PyPI rejects us.
-    ## For multiple authors, they tell us to get a mailing list :/
-    author = "See homepage for a complete list of contributors",
-    author_email = " ",
-
-    url = "https://github.com/MicronOxford/microscope",
-
-    packages = setuptools.find_packages(),
-
-    python_requires = '>=3.5',
-
-    install_requires = [
-        "Pyro4",
-        "hidapi",
-        "numpy",
-        "pyserial",
-    ],
-
-    entry_points = {
-        'console_scripts' : [
-            'deviceserver = microscope.deviceserver:__main__',
+    name=project_name,
+    version=project_version,
+    description="An extensible microscope hardware interface.",
+    long_description=open("README", "r").read(),
+    license="GPL-3.0+",
+    # We need an author and an author_email value or PyPI rejects us.
+    # For multiple authors, they tell us to get a mailing list :/
+    author="See homepage for a complete list of contributors",
+    author_email=" ",
+    url="https://github.com/MicronOxford/microscope",
+    packages=setuptools.find_packages(),
+    python_requires=">=3.6",
+    install_requires=["Pillow", "Pyro4", "hidapi", "numpy", "pyserial",],
+    extras_require={"GUI": ["PySide2"],},
+    entry_points={
+        "console_scripts": [
+            "device-server = microscope.device_server:__main__",
+            "deviceserver = microscope.device_server:__main__",
+            "microscope-gui = microscope.gui:_setuptools_entry_point [GUI]",
         ]
     },
-
-    ## https://pypi.python.org/pypi?:action=list_classifiers
-    classifiers = [
+    # https://pypi.python.org/pypi?:action=list_classifiers
+    classifiers=[
         "Intended Audience :: Science/Research",
         "Topic :: Scientific/Engineering",
         "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)",
     ],
-    test_suite = "microscope.testsuite",
-
-    command_options = {
-        'build_sphinx' : {
-            ## The dict for command_options must be of the form
-            ## '(option, (source, value))' where source is the
-            ## filename where that information came from.
-            'project': ('setup.py', project_name),
-            'version': ('setup.py', project_version),
-            'release': ('setup.py', project_version),
-            'source_dir' : ('setup.py', 'doc'),
+    test_suite="microscope.testsuite",
+    command_options={
+        "build_sphinx": {
+            # The dict for command_options must be of the form
+            # '(option, (source, value))' where source is the
+            # filename where that information came from.
+            "project": ("setup.py", project_name),
+            "version": ("setup.py", project_version),
+            "release": ("setup.py", project_version),
+            "source_dir": ("setup.py", "doc"),
         },
     },
-
-    cmdclass = {
-        'build_sphinx' : BuildDoc,
-        'sdist' : sdist,
-    },
+    cmdclass={"build_sphinx": BuildDoc, "sdist": sdist,},
 )
