@@ -1,7 +1,8 @@
-#!/usr/bin/env python
-# -*- coding: utf-8 -*-
+#!/usr/bin/env python3
 
-## Copyright (C) 2017 David Pinto <david.pinto@bioch.ox.ac.uk>
+## Copyright (C) 2020 David Miguel Susano Pinto <carandraug@gmail.com>
+##
+## This file is part of Microscope.
 ##
 ## Microscope is free software: you can redistribute it and/or modify
 ## it under the terms of the GNU General Public License as published by
@@ -21,33 +22,29 @@
 
 import ctypes
 import os
-
-from ctypes import c_char_p
-from ctypes import c_double
-from ctypes import c_int
-from ctypes import c_size_t
-from ctypes import c_uint32
-
+from ctypes import c_char_p, c_double, c_int, c_size_t, c_uint32
 
 if os.name in ("nt", "ce"):
     SDK = ctypes.WinDLL("ASDK")
 else:
-    ## Not actually tested yet
+    # Not actually tested yet
     SDK = ctypes.CDLL("libasdk.so")
 
 
 class DM(ctypes.Structure):
     pass
+
+
 pDM = ctypes.POINTER(DM)
 
-## We have this "typedefs" to ease matching with alpao's headers.
+# We have this "typedefs" to ease matching with alpao's headers.
 CStr = c_char_p
 Scalar = c_double
 Scalar_p = ctypes.POINTER(Scalar)
 UInt = c_uint32
 Size_T = c_size_t
 
-COMPL_STAT = c_int # enum for function completion status
+COMPL_STAT = c_int  # enum for function completion status
 SUCCESS = 0
 FAILURE = -1
 
@@ -58,10 +55,12 @@ def make_prototype(name, argtypes, restype=COMPL_STAT):
     func.restype = restype
     return func
 
+
 Get = make_prototype("asdkGet", [pDM, CStr, Scalar_p])
 
-GetLastError = make_prototype("asdkGetLastError",
-                              [ctypes.POINTER(UInt), CStr, Size_T])
+GetLastError = make_prototype(
+    "asdkGetLastError", [ctypes.POINTER(UInt), CStr, Size_T]
+)
 
 Init = make_prototype("asdkInit", [CStr], pDM)
 
@@ -72,3 +71,5 @@ Send = make_prototype("asdkSend", [pDM, Scalar_p])
 SendPattern = make_prototype("asdkSendPattern", [pDM, Scalar_p, UInt, UInt])
 
 Set = make_prototype("asdkSet", [pDM, CStr, Scalar])
+
+Stop = make_prototype("asdkStop", [pDM])
