@@ -1,20 +1,22 @@
-#!/usr/bin/python
-# -*- coding: utf-8
-#
-# Copyright 2017-2019 Mick Phillips (mick.phillips@gmail.com)
-#
-# This program is free software: you can redistribute it and/or modify
-# it under the terms of the GNU General Public License as published by
-# the Free Software Foundation, either version 3 of the License, or
-# (at your option) any later version.
-#
-# This program is distributed in the hope that it will be useful,
-# but WITHOUT ANY WARRANTY; without even the implied warranty of
-# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-# GNU General Public License for more details.
-#
-# You should have received a copy of the GNU General Public License
-# along with this program.  If not, see <http://www.gnu.org/licenses/>.
+#!/usr/bin/env python3
+
+## Copyright (C) 2020 Mick Phillips <mick.phillips@gmail.com>
+##
+## This file is part of Microscope.
+##
+## Microscope is free software: you can redistribute it and/or modify
+## it under the terms of the GNU General Public License as published by
+## the Free Software Foundation, either version 3 of the License, or
+## (at your option) any later version.
+##
+## Microscope is distributed in the hope that it will be useful,
+## but WITHOUT ANY WARRANTY; without even the implied warranty of
+## MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+## GNU General Public License for more details.
+##
+## You should have received a copy of the GNU General Public License
+## along with Microscope.  If not, see <http://www.gnu.org/licenses/>.
+
 """pvcam library wrapper.
 
 This module exposes pvcam C library functions in python.
@@ -132,47 +134,43 @@ import logging
 import os
 import platform
 import time
+import weakref
 
-import Pyro4
 import numpy as np
+import Pyro4
 
-from microscope import devices
-from microscope.devices import ROI, Binning
-from microscope.devices import keep_acquiring
-
+import microscope
+import microscope.abc
 
 _logger = logging.getLogger(__name__)
 
 
 # Readout transform mapping - {CHIP_NAME: {port: transform}}
-READOUT_TRANSFORMS = {
-    'Evolve-5': {0: (0,0,0),
-                 1: (1,0,0)}
-}
+READOUT_TRANSFORMS = {"Evolve-5": {0: (0, 0, 0), 1: (1, 0, 0)}}
 
 # === Data types ===
 # Base typedefs, from pvcam SDK master.h
-#typedef unsigned short rs_bool;
+# typedef unsigned short rs_bool;
 rs_bool = ctypes.c_ushort
-#typedef signed char    int8;
+# typedef signed char    int8;
 int8 = ctypes.c_byte
-#typedef unsigned char  uns8;
+# typedef unsigned char  uns8;
 uns8 = ctypes.c_ubyte
-#typedef short          int16;
+# typedef short          int16;
 int16 = ctypes.c_short
-#typedef unsigned short uns16;
+# typedef unsigned short uns16;
 uns16 = ctypes.c_ushort
-#typedef int            int32;
+# typedef int            int32;
 int32 = ctypes.c_int32
-#typedef unsigned int   uns32;
+# typedef unsigned int   uns32;
 uns32 = ctypes.c_uint32
-#typedef float          flt32;
+# typedef float          flt32;
 flt32 = ctypes.c_float
-#typedef double         flt64;
+# typedef double         flt64;
 flt64 = ctypes.c_double
-#typedef unsigned long long ulong64;
+# typedef unsigned long long ulong64;
 ulong64 = ctypes.c_ulonglong
-#typedef signed long long long64;
+# typedef signed long long long64;
 long64 = ctypes.c_longlong
 # enums
 enumtype = ctypes.c_int32
@@ -504,42 +502,54 @@ PARAM_LAST_MUXED_SIGNAL = 84082869
 # === C structures ===
 # GUID for #FRAME_INFO structure.
 class PVCAM_FRAME_INFO_GUID(ctypes.Structure):
-    _fields_ = [("f1", uns32),
-                ("f2", uns16),
-                ("f3", uns16),
-                ("f4", uns8 * 8),]
+    _fields_ = [
+        ("f1", uns32),
+        ("f2", uns16),
+        ("f3", uns16),
+        ("f4", uns8 * 8),
+    ]
+
 
 # Structure used to uniquely identify frames in the camera.
 class FRAME_INFO(ctypes.Structure):
-    _fields_ = [("FrameInfoGUID", PVCAM_FRAME_INFO_GUID),
-                ("hCam", int16),
-                ("FrameNr", int32),
-                ("TimeStamp", long64),
-                ("ReadoutTime", int32),
-                ("TimeStampBOF", long64),]
+    _fields_ = [
+        ("FrameInfoGUID", PVCAM_FRAME_INFO_GUID),
+        ("hCam", int16),
+        ("FrameNr", int32),
+        ("TimeStamp", long64),
+        ("ReadoutTime", int32),
+        ("TimeStampBOF", long64),
+    ]
 
 
 class smart_stream_type(ctypes.Structure):
-    _fields_ = [("entries", uns16),
-                ("params", uns32),]
+    _fields_ = [
+        ("entries", uns16),
+        ("params", uns32),
+    ]
 
 
 class rgn_type(ctypes.Structure):
-    _fields_ = [("s1", uns16),
-               ("s2", uns16),
-               ("sbin", uns16),
-               ("p1", uns16),
-               ("p2", uns16),
-               ("pbin", uns16),]
+    _fields_ = [
+        ("s1", uns16),
+        ("s2", uns16),
+        ("sbin", uns16),
+        ("p1", uns16),
+        ("p2", uns16),
+        ("pbin", uns16),
+    ]
 
 
 class io_struct(ctypes.Structure):
     pass
 
-io_struct._fields_ = [("io_port", uns16),
-                     ("io_type", uns32),
-                     ("state", flt64),
-                     ("next", ctypes.POINTER(io_struct))]
+
+io_struct._fields_ = [
+    ("io_port", uns16),
+    ("io_type", uns32),
+    ("state", flt64),
+    ("next", ctypes.POINTER(io_struct)),
+]
 
 
 class io_list(ctypes.Structure):
@@ -555,6 +565,7 @@ class io_list(ctypes.Structure):
         ("pre_close", ctypes.POINTER(io_struct)),
         ("post_close", ctypes.POINTER(io_struct)),
     ]
+
 
 class active_camera_type(ctypes.Structure):
     _fields_ = [
@@ -584,7 +595,7 @@ class active_camera_type(ctypes.Structure):
 class md_frame_header(ctypes.Structure):
     _fields_ = [
         ("signature", uns32),
-        ("version", uns8 ),
+        ("version", uns8),
         ("frameNr", uns32),
         ("roiCount", uns16),
         ("timestampBOF", uns32),
@@ -597,7 +608,8 @@ class md_frame_header(ctypes.Structure):
         ("colorMask", uns8),
         ("flags", uns8),
         ("extendedMdSize", uns16),
-        ("_reserved", uns8*8),]
+        ("_reserved", uns8 * 8),
+    ]
 
 
 class md_frame_roi_header(ctypes.Structure):
@@ -608,8 +620,9 @@ class md_frame_roi_header(ctypes.Structure):
         ("roi", rgn_type),
         ("flags", uns8),
         ("extendedMdSize", uns16),
-        ("_reserved", uns8*7),
+        ("_reserved", uns8 * 7),
     ]
+
 
 class md_ext_item_info(ctypes.Structure):
     _fields_ = [
@@ -618,19 +631,21 @@ class md_ext_item_info(ctypes.Structure):
         ("name", ctypes.c_char_p),
     ]
 
+
 class md_ext_item(ctypes.Structure):
     _fields_ = [
-        ("tagInfo", ctypes.POINTER(md_ext_item_info)), #
-        ("value", ctypes.c_void_p)
+        ("tagInfo", ctypes.POINTER(md_ext_item_info)),  #
+        ("value", ctypes.c_void_p),
     ]
 
 
 class md_ext_item_collection(ctypes.Structure):
     _fields_ = [
-        ("list", md_ext_item*PL_MD_EXT_TAGS_MAX_SUPPORTED),
-        ("map", ctypes.POINTER(md_ext_item)*PL_MD_EXT_TAGS_MAX_SUPPORTED),
+        ("list", md_ext_item * PL_MD_EXT_TAGS_MAX_SUPPORTED),
+        ("map", ctypes.POINTER(md_ext_item) * PL_MD_EXT_TAGS_MAX_SUPPORTED),
         ("count", uns16),
     ]
+
 
 class md_frame_roi(ctypes.Structure):
     _fields_ = [
@@ -640,6 +655,7 @@ class md_frame_roi(ctypes.Structure):
         ("extMdData", ctypes.c_void_p),
         ("extMdDataSize", uns16),
     ]
+
 
 class md_frame(ctypes.Structure):
     _fields_ = [
@@ -653,13 +669,13 @@ class md_frame(ctypes.Structure):
     ]
 
 
-if os.name in ('nt', 'ce'):
-    if platform.architecture()[0] == '32bit':
-        _lib = ctypes.WinDLL('pvcam32')
+if os.name in ("nt", "ce"):
+    if platform.architecture()[0] == "32bit":
+        _lib = ctypes.WinDLL("pvcam32")
     else:
-        _lib = ctypes.WinDLL('pvcam64')
+        _lib = ctypes.WinDLL("pvcam64")
 else:
-    _lib = ctypes.CDLL('pvcam.so')
+    _lib = ctypes.CDLL("pvcam.so")
 
 ### Functions ###
 STRING = ctypes.c_char_p
@@ -704,6 +720,7 @@ def stripMeta(val):
     else:
         return val
 
+
 # Function type for callbacks.
 CALLBACK = ctypes.CFUNCTYPE(ctypes.c_void_p)
 
@@ -712,6 +729,7 @@ class dllFunction:
     """Expose a DLL function to python.
 
     (Again, largely nicked from PYME.)"""
+
     def __init__(self, name, args=[], argnames=[], buf_len=-1, lib=_lib):
         self.f = getattr(lib, name)
         self.f.restype = rs_bool
@@ -727,12 +745,12 @@ class dllFunction:
 
         self.buf_len = buf_len
 
-        docstring = name + '\n\nArguments:\n===========\n'
+        docstring = name + "\n\nArguments:\n===========\n"
         for i in range(len(args)):
-            an = ''
+            an = ""
             if i < len(argnames):
                 an = argnames[i]
-            docstring += '\t%s\t%s\n' % (args[i], an)
+            docstring += "\t%s\t%s\n" % (args[i], an)
 
         self.f.__doc__ = docstring
 
@@ -742,9 +760,9 @@ class dllFunction:
         ret = []
         # pl_get_param buffer length depends on the parameter being fetched, so
         # use kwargs to pass buffer length.
-        if 'buf_len' in kwargs:
-            bs = kwargs['buf_len']
-        elif self.name == 'pl_get_enum_param':
+        if "buf_len" in kwargs:
+            bs = kwargs["buf_len"]
+        elif self.name == "pl_get_enum_param":
             # last argument is buffer length
             bs = args[-1]
         elif self.buf_len >= 0:
@@ -757,7 +775,9 @@ class dllFunction:
 
         for j in range(len(self.inp)):
             if self.inp[j]:  # an input
-                if self.f.argtypes[j] is CALLBACK and not isinstance(args[i], CALLBACK):
+                if self.f.argtypes[j] is CALLBACK and not isinstance(
+                    args[i], CALLBACK
+                ):
                     ars.append(CALLBACK(args[i]))
                 else:
                     ars.append(args[i])
@@ -776,7 +796,9 @@ class dllFunction:
             err_code = _lib.pl_error_code()
             err_msg = ctypes.create_string_buffer(ERROR_MSG_LEN)
             _lib.pl_error_message(err_code, err_msg)
-            raise Exception('pvcam error %d: %s' % (err_code, err_msg.value))
+            raise microscope.DeviceError(
+                "pvcam error %d: %s" % (err_code, err_msg.value)
+            )
 
         if len(ret) == 0:
             return None
@@ -801,85 +823,179 @@ def dllFunc(name, args=[], argnames=[], buf_len=0):
 
 """DLL function imports."""
 # Class 0 functions - library
-dllFunc('pl_pvcam_get_ver', [OUTPUT(uns16)], ['version'])
-dllFunc('pl_pvcam_init')
-dllFunc('pl_pvcam_uninit')
+dllFunc("pl_pvcam_get_ver", [OUTPUT(uns16)], ["version"])
+dllFunc("pl_pvcam_init")
+dllFunc("pl_pvcam_uninit")
 # Class 0 functions - camera
-dllFunc('pl_cam_close', [int16], ['hcam'])
-dllFunc('pl_cam_get_name',
-        [int16, OUTSTRING],
-        ['can_num', 'cam_name'], buf_len=CAM_NAME_LEN)
-dllFunc('pl_cam_get_total', [OUTPUT(int16),], ['total_cams',])
-dllFunc('pl_cam_open',
-        [STRING, OUTPUT(int16), int16],
-        ['cam_name', 'hcam', 'o_mode'])
-dllFunc('pl_cam_register_callback',
-        [int16, int32, CALLBACK],
-        ['hcam', 'event', 'Callback'])
-dllFunc('pl_cam_register_callback_ex',
-        [int16, int32, CALLBACK, ctypes.c_void_p],
-        ['hcam', 'event', 'Callback', 'Context'])
-dllFunc('pl_cam_register_callback_ex2',
-        [int16, int32, CALLBACK],
-        ['hcam', 'event', 'Callback'])
-dllFunc('pl_cam_register_callback_ex3',
-        [int16, int32, CALLBACK, ctypes.c_void_p],
-        ['hcam', 'event', 'Callback', 'Context'])
-dllFunc('pl_cam_deregister_callback',
-        [int16, ctypes.c_void_p],
-        ['hcam', 'event'])
+dllFunc("pl_cam_close", [int16], ["hcam"])
+dllFunc(
+    "pl_cam_get_name",
+    [int16, OUTSTRING],
+    ["can_num", "cam_name"],
+    buf_len=CAM_NAME_LEN,
+)
+dllFunc("pl_cam_get_total", [OUTPUT(int16),], ["total_cams",])
+dllFunc(
+    "pl_cam_open",
+    [STRING, OUTPUT(int16), int16],
+    ["cam_name", "hcam", "o_mode"],
+)
+dllFunc(
+    "pl_cam_register_callback",
+    [int16, int32, CALLBACK],
+    ["hcam", "event", "Callback"],
+)
+dllFunc(
+    "pl_cam_register_callback_ex",
+    [int16, int32, CALLBACK, ctypes.c_void_p],
+    ["hcam", "event", "Callback", "Context"],
+)
+dllFunc(
+    "pl_cam_register_callback_ex2",
+    [int16, int32, CALLBACK],
+    ["hcam", "event", "Callback"],
+)
+dllFunc(
+    "pl_cam_register_callback_ex3",
+    [int16, int32, CALLBACK, ctypes.c_void_p],
+    ["hcam", "event", "Callback", "Context"],
+)
+dllFunc(
+    "pl_cam_deregister_callback", [int16, ctypes.c_void_p], ["hcam", "event"]
+)
 # Class 1 functions - error handling. Handled in dllFunction.
 # Class 2 functions - configuration/setup.
-dllFunc('pl_get_param', [int16, uns32, int16, OUTPUT(ctypes.c_void_p)],
-        ['hcam', 'param_id', 'param_attrib', 'param_value'])
-dllFunc('pl_set_param', [int16, uns32, ctypes.c_void_p],
-        ['hcam', 'param_id', 'param_value'])
-dllFunc('pl_get_enum_param',
-        [int16, uns32, uns32, OUTPUT(int32), OUTSTRING, uns32],
-        ['hcam', 'param_id', 'index', 'value', 'desc', 'length'])
-dllFunc('pl_enum_str_length', [int16, uns32, uns32, OUTPUT(uns32)],
-        ['hcam', 'param_id', 'index', 'length'])
-dllFunc('pl_pp_reset', [int16,], ['hcam'])
-dllFunc('pl_create_smart_stream_struct', [OUTPUT(smart_stream_type), uns16],
-        ['pSmtStruct', 'entries'])
-dllFunc('pl_release_smart_stream_struct', [ctypes.POINTER(smart_stream_type),],
-        ['pSmtStruct',])
-dllFunc('pl_create_frame_info_struct', [OUTPUT(FRAME_INFO),],
-        ['pNewFrameInfo'])
-dllFunc('pl_release_frame_info_struct', [ctypes.POINTER(FRAME_INFO),],
-        ['pFrameInfoToDel',])
-dllFunc('pl_exp_abort', [int16, int16], ['hcam', 'cam_state'])
-dllFunc('pl_exp_setup_seq',
-        [int16, uns16, uns16, ctypes.POINTER(rgn_type), int16, uns32, OUTPUT(uns32)],
-        ['hcam', 'exp_total', 'rgn_total', 'rgn_array', 'exp_mode', 'exposure_time', 'exp_bytes'])
-dllFunc('pl_exp_start_seq', [int16, ctypes.c_void_p], ['hcam', 'pixel_stream'])
-dllFunc('pl_exp_setup_cont',
-        [int16, uns16, ctypes.POINTER(rgn_type), int16, uns32, OUTPUT(uns32), int16],
-        ['hcam', 'rgn_total', 'rgn_array', 'exp_mode', 'exposure_time', 'exp_bytes', 'buffer_mode'])
-dllFunc('pl_exp_start_cont',
-        [int16, ctypes.c_void_p, uns32],
-        ['hcam', 'pixel_stream', 'size'])
-dllFunc('pl_exp_check_status',
-        [int16, OUTPUT(int16), OUTPUT(uns32)],
-        ['hcam', 'status', 'bytes_arrived'])
-dllFunc('pl_exp_check_cont_status',
-        [int16, OUTPUT(int16), OUTPUT(uns32), OUTPUT(uns32)],
-        ['hcam', 'status', 'bytes_arrived', 'buffer_cnt'])
-dllFunc('pl_exp_check_cont_status_ex',
-        [int16, OUTPUT(int16), OUTPUT(uns32), OUTPUT(uns32), ctypes.POINTER(FRAME_INFO)],
-        ['hcam', 'status', 'byte_cnt', 'buffer_cnt', 'pFrameInfo'])
-dllFunc('pl_exp_get_latest_frame', [int16, OUTPUT(ctypes.c_void_p)], ['hcam', 'frame'])
-dllFunc('pl_exp_get_latest_frame_ex',
-        [int16, OUTPUT(ctypes.c_void_p), ctypes.POINTER(FRAME_INFO)],
-        ['hcam', 'frame', 'pFrameInfo'])
-dllFunc('pl_exp_get_oldest_frame', [int16, OUTPUT(ctypes.c_void_p)], ['hcam', 'frame'])
-dllFunc('pl_exp_get_oldest_frame_ex',
-        [int16, OUTPUT(ctypes.c_void_p), ctypes.POINTER(FRAME_INFO)],
-        ['hcam', 'frame', 'pFrameInfo'])
-dllFunc('pl_exp_unlock_oldest_frame', [int16], ['hcam'])
-dllFunc('pl_exp_stop_cont', [int16, int16], ['hcam', 'cam_state'])
-dllFunc('pl_exp_abort', [int16, int16], ['hcam', 'cam_state'])
-dllFunc('pl_exp_finish_seq', [int16, ctypes.c_void_p], ['hcam', 'pixel_stream'])
+dllFunc(
+    "pl_get_param",
+    [int16, uns32, int16, OUTPUT(ctypes.c_void_p)],
+    ["hcam", "param_id", "param_attrib", "param_value"],
+)
+dllFunc(
+    "pl_set_param",
+    [int16, uns32, ctypes.c_void_p],
+    ["hcam", "param_id", "param_value"],
+)
+dllFunc(
+    "pl_get_enum_param",
+    [int16, uns32, uns32, OUTPUT(int32), OUTSTRING, uns32],
+    ["hcam", "param_id", "index", "value", "desc", "length"],
+)
+dllFunc(
+    "pl_enum_str_length",
+    [int16, uns32, uns32, OUTPUT(uns32)],
+    ["hcam", "param_id", "index", "length"],
+)
+dllFunc("pl_pp_reset", [int16,], ["hcam"])
+dllFunc(
+    "pl_create_smart_stream_struct",
+    [OUTPUT(smart_stream_type), uns16],
+    ["pSmtStruct", "entries"],
+)
+dllFunc(
+    "pl_release_smart_stream_struct",
+    [ctypes.POINTER(smart_stream_type),],
+    ["pSmtStruct",],
+)
+dllFunc("pl_create_frame_info_struct", [OUTPUT(FRAME_INFO),], ["pNewFrameInfo"])
+dllFunc(
+    "pl_release_frame_info_struct",
+    [ctypes.POINTER(FRAME_INFO),],
+    ["pFrameInfoToDel",],
+)
+dllFunc("pl_exp_abort", [int16, int16], ["hcam", "cam_state"])
+dllFunc(
+    "pl_exp_setup_seq",
+    [
+        int16,
+        uns16,
+        uns16,
+        ctypes.POINTER(rgn_type),
+        int16,
+        uns32,
+        OUTPUT(uns32),
+    ],
+    [
+        "hcam",
+        "exp_total",
+        "rgn_total",
+        "rgn_array",
+        "exp_mode",
+        "exposure_time",
+        "exp_bytes",
+    ],
+)
+dllFunc("pl_exp_start_seq", [int16, ctypes.c_void_p], ["hcam", "pixel_stream"])
+dllFunc(
+    "pl_exp_setup_cont",
+    [
+        int16,
+        uns16,
+        ctypes.POINTER(rgn_type),
+        int16,
+        uns32,
+        OUTPUT(uns32),
+        int16,
+    ],
+    [
+        "hcam",
+        "rgn_total",
+        "rgn_array",
+        "exp_mode",
+        "exposure_time",
+        "exp_bytes",
+        "buffer_mode",
+    ],
+)
+dllFunc(
+    "pl_exp_start_cont",
+    [int16, ctypes.c_void_p, uns32],
+    ["hcam", "pixel_stream", "size"],
+)
+dllFunc(
+    "pl_exp_check_status",
+    [int16, OUTPUT(int16), OUTPUT(uns32)],
+    ["hcam", "status", "bytes_arrived"],
+)
+dllFunc(
+    "pl_exp_check_cont_status",
+    [int16, OUTPUT(int16), OUTPUT(uns32), OUTPUT(uns32)],
+    ["hcam", "status", "bytes_arrived", "buffer_cnt"],
+)
+dllFunc(
+    "pl_exp_check_cont_status_ex",
+    [
+        int16,
+        OUTPUT(int16),
+        OUTPUT(uns32),
+        OUTPUT(uns32),
+        ctypes.POINTER(FRAME_INFO),
+    ],
+    ["hcam", "status", "byte_cnt", "buffer_cnt", "pFrameInfo"],
+)
+dllFunc(
+    "pl_exp_get_latest_frame",
+    [int16, OUTPUT(ctypes.c_void_p)],
+    ["hcam", "frame"],
+)
+dllFunc(
+    "pl_exp_get_latest_frame_ex",
+    [int16, OUTPUT(ctypes.c_void_p), ctypes.POINTER(FRAME_INFO)],
+    ["hcam", "frame", "pFrameInfo"],
+)
+dllFunc(
+    "pl_exp_get_oldest_frame",
+    [int16, OUTPUT(ctypes.c_void_p)],
+    ["hcam", "frame"],
+)
+dllFunc(
+    "pl_exp_get_oldest_frame_ex",
+    [int16, OUTPUT(ctypes.c_void_p), ctypes.POINTER(FRAME_INFO)],
+    ["hcam", "frame", "pFrameInfo"],
+)
+dllFunc("pl_exp_unlock_oldest_frame", [int16], ["hcam"])
+dllFunc("pl_exp_stop_cont", [int16, int16], ["hcam", "cam_state"])
+dllFunc("pl_exp_abort", [int16, int16], ["hcam", "cam_state"])
+dllFunc("pl_exp_finish_seq", [int16, ctypes.c_void_p], ["hcam", "pixel_stream"])
 
 
 # Map ATTR_ enums to the return type for that ATTR.
@@ -904,7 +1020,7 @@ _typemap = {
     TYPE_UNS16: uns16,
     TYPE_UNS32: uns32,
     TYPE_UNS64: ulong64,
-    TYPE_ENUM: int32, # from SDK documentation
+    TYPE_ENUM: int32,  # from SDK documentation
     TYPE_BOOLEAN: rs_bool,
     TYPE_INT8: int8,
     TYPE_CHAR_PTR: ctypes.c_char_p,
@@ -913,28 +1029,29 @@ _typemap = {
     TYPE_INT64: long64,
     TYPE_SMART_STREAM_TYPE: smart_stream_type,
     TYPE_SMART_STREAM_TYPE_PTR: ctypes.POINTER(smart_stream_type),
-    TYPE_FLT32: flt32,}
+    TYPE_FLT32: flt32,
+}
 
 
 # Map TYPE enums to the appropriate setting dtype.
 _dtypemap = {
-    TYPE_INT16: 'int',
-    TYPE_INT32: 'int',
-    TYPE_FLT64: 'float',
-    TYPE_UNS8: 'int',
-    TYPE_UNS16: 'int',
-    TYPE_UNS32: 'int',
-    TYPE_UNS64: 'int',
-    TYPE_ENUM: 'enum',
-    TYPE_BOOLEAN: 'bool',
-    TYPE_INT8: 'int',
-    TYPE_CHAR_PTR: 'str',
+    TYPE_INT16: "int",
+    TYPE_INT32: "int",
+    TYPE_FLT64: "float",
+    TYPE_UNS8: "int",
+    TYPE_UNS16: "int",
+    TYPE_UNS32: "int",
+    TYPE_UNS64: "int",
+    TYPE_ENUM: "enum",
+    TYPE_BOOLEAN: "bool",
+    TYPE_INT8: "int",
+    TYPE_CHAR_PTR: "str",
     TYPE_VOID_PTR: None,
     TYPE_VOID_PTR_PTR: None,
-    TYPE_INT64: 'int',
+    TYPE_INT64: "int",
     TYPE_SMART_STREAM_TYPE: None,
     TYPE_SMART_STREAM_TYPE_PTR: None,
-    TYPE_FLT32: 'float',
+    TYPE_FLT32: "float",
 }
 
 # Mapping of param ids to maximum string lengths.
@@ -955,8 +1072,11 @@ _length_map = {
 }
 
 # map PARAM enums to the parameter name
-_param_to_name = {globals()[param]:param for param in globals()
-                  if (param.startswith('PARAM_') and param != 'PARAM_NAME_LEN')}
+_param_to_name = {
+    globals()[param]: param
+    for param in globals()
+    if (param.startswith("PARAM_") and param != "PARAM_NAME_LEN")
+}
 
 
 def get_param_type(param_id):
@@ -972,18 +1092,21 @@ def get_param_dtype(param_id):
 
 
 # Map status codes to strings.
-STATUS_STRINGS = {READOUT_NOT_ACTIVE: 'READOUT_NOT_ACTIVE',
-                  EXPOSURE_IN_PROGRESS: 'EXPOSURE_IN_PROGRESS',
-                  READOUT_IN_PROGRESS: 'READOUT_IN_PROGRESS',
-                  READOUT_COMPLETE: 'READOUT_COMPLETE',
-                  READOUT_FAILED: 'READOUT_FAILED',
-                  FRAME_AVAILABLE: 'FRAME_AVAILABLE',}
+STATUS_STRINGS = {
+    READOUT_NOT_ACTIVE: "READOUT_NOT_ACTIVE",
+    EXPOSURE_IN_PROGRESS: "EXPOSURE_IN_PROGRESS",
+    READOUT_IN_PROGRESS: "READOUT_IN_PROGRESS",
+    READOUT_COMPLETE: "READOUT_COMPLETE",
+    READOUT_FAILED: "READOUT_FAILED",
+    FRAME_AVAILABLE: "FRAME_AVAILABLE",
+}
 
 
 # === Python classes ===
 # Trigger modes.
 class TriggerMode:
     """A microscope trigger mode using PVCAM PMODES."""
+
     def __init__(self, id, label, pv_mode, microscope_mode):
         self.id = id
         self.label = label
@@ -995,41 +1118,66 @@ class TriggerMode:
 
 
 # Enumerate trigger types.
-(TRIG_SOFT, TRIG_TIMED, TRIG_VARIABLE, TRIG_FIRST, TRIG_STROBED, TRIG_BULB) = range(6)
+(
+    TRIG_SOFT,
+    TRIG_TIMED,
+    TRIG_VARIABLE,
+    TRIG_FIRST,
+    TRIG_STROBED,
+    TRIG_BULB,
+) = range(6)
 
 # Trigger mode definitions.
 TRIGGER_MODES = {
-    TRIG_SOFT: TriggerMode(TRIG_SOFT, 'software', TIMED_MODE, devices.TRIGGER_SOFT),
-    TRIG_TIMED: TriggerMode(TRIG_TIMED, 'timed', TIMED_MODE, -1),
-    TRIG_VARIABLE: TriggerMode(TRIG_VARIABLE, 'variable timed', VARIABLE_TIMED_MODE, -1),
-    TRIG_FIRST: TriggerMode(TRIG_FIRST, 'trig. first', TRIGGER_FIRST_MODE, devices.TRIGGER_BEFORE),
-    TRIG_STROBED: TriggerMode(TRIG_STROBED, 'strobed', STROBED_MODE, devices.TRIGGER_BEFORE),
-    TRIG_BULB: TriggerMode(TRIG_BULB, 'bulb', BULB_MODE, devices.TRIGGER_DURATION),
+    TRIG_SOFT: TriggerMode(
+        TRIG_SOFT, "software", TIMED_MODE, microscope.abc.TRIGGER_SOFT
+    ),
+    TRIG_TIMED: TriggerMode(TRIG_TIMED, "timed", TIMED_MODE, -1),
+    TRIG_VARIABLE: TriggerMode(
+        TRIG_VARIABLE, "variable timed", VARIABLE_TIMED_MODE, -1
+    ),
+    TRIG_FIRST: TriggerMode(
+        TRIG_FIRST,
+        "trig. first",
+        TRIGGER_FIRST_MODE,
+        microscope.abc.TRIGGER_BEFORE,
+    ),
+    TRIG_STROBED: TriggerMode(
+        TRIG_STROBED, "strobed", STROBED_MODE, microscope.abc.TRIGGER_BEFORE
+    ),
+    TRIG_BULB: TriggerMode(
+        TRIG_BULB, "bulb", BULB_MODE, microscope.abc.TRIGGER_DURATION
+    ),
 }
 
-class PVParam():
+
+class PVParam:
     """A wrapper around PVCAM parameters."""
+
     @staticmethod
     def factory(camera, param_id):
         """Create a PVParam or appropriate subclass"""
         # A mapping of pv parameters types to python types.
         #   None means unsupported.
         #   Parameters omitted from the mapping will default to PVParam.
-        __types__ = {TYPE_SMART_STREAM_TYPE: None,
-                     TYPE_SMART_STREAM_TYPE_PTR: None,
-                     TYPE_VOID_PTR: None,
-                     TYPE_VOID_PTR_PTR: None,
-                     TYPE_ENUM: PVEnumParam,
-                     TYPE_CHAR_PTR: PVStringParam}
+        __types__ = {
+            TYPE_SMART_STREAM_TYPE: None,
+            TYPE_SMART_STREAM_TYPE_PTR: None,
+            TYPE_VOID_PTR: None,
+            TYPE_VOID_PTR_PTR: None,
+            TYPE_ENUM: PVEnumParam,
+            TYPE_CHAR_PTR: PVStringParam,
+        }
         # Determine the appropiate type from its id.
         pvtype = __types__.get(param_id >> 24 & 255, PVParam)
         if pvtype is None:
-            raise Exception("Parameter %s not supported" % _param_to_name[param_id])
+            raise microscope.LibraryLoadError(
+                "Parameter %s not supported" % _param_to_name[param_id]
+            )
         return pvtype(camera, param_id)
 
     def __init__(self, camera, param_id):
         # Use a weakref back to the camera to avoid circular dependency.
-        import weakref
         self.cam = weakref.proxy(camera)
         self.param_id = param_id
         self.name = _param_to_name[param_id]
@@ -1047,7 +1195,7 @@ class PVParam():
             ref = ctypes.byref(new_value)
         except TypeError:
             # Need to convert python type to ctype first.
-            ref = ctypes.byref(self._ctype(new_value) )
+            ref = ctypes.byref(self._ctype(new_value))
         _set_param(self.cam.handle, self.param_id, ref)
         # Read back the value to update cache.
         self._query(force_query=True)
@@ -1057,22 +1205,28 @@ class PVParam():
 
         This returns pythonic types, not ctypes."""
         err = None
-        key = (self, what) # key for cache
+        key = (self, what)  # key for cache
         if self.cam._acquiring and not force_query:
             return self.__cache.get(key, None)
         if what == ATTR_AVAIL:
             return self.available
         elif not self.available:
-            raise Exception("Parameter %s is not available" % self.name)
-        rtype = _attr_map[what] # return type
+            raise microscope.DeviceError(
+                "Parameter %s is not available" % self.name
+            )
+        rtype = _attr_map[what]  # return type
         if not rtype:
             rtype = _get_param(self.cam.handle, self.param_id, ATTR_TYPE)
         if rtype.value == TYPE_CHAR_PTR:
             buf_len = _length_map[self.param_id]
             if not buf_len:
-                raise Exception('pvcam: parameter %s not supported in python.' % self.name)
+                raise microscope.DeviceError(
+                    "pvcam: parameter %s not supported in python." % self.name
+                )
             try:
-                result = _get_param(self.cam.handle, self.param_id, what, buf_len=buf_len)
+                result = _get_param(
+                    self.cam.handle, self.param_id, what, buf_len=buf_len
+                )
             except Exception as e:
                 err = e
             else:
@@ -1085,9 +1239,10 @@ class PVParam():
             else:
                 result = ctypes.POINTER(self._ctype)(result).contents.value
         # Test on err.args prevents indexing into empty tuple.
-        if err and err.args and err.args[0].startswith('pvcam error 49'):
-            _logger.warn("Parameter %s not available due to camera state.",
-                         self.name)
+        if err and err.args and err.args[0].startswith("pvcam error 49"):
+            _logger.warn(
+                "Parameter %s not available due to camera state.", self.name
+            )
             result = None
         elif err:
             raise e
@@ -1098,12 +1253,16 @@ class PVParam():
     @property
     def access(self):
         """Return parameter access attribute."""
-        return int(_get_param(self.cam.handle, self.param_id, ATTR_ACCESS).value)
+        return int(
+            _get_param(self.cam.handle, self.param_id, ATTR_ACCESS).value
+        )
 
     @property
     def available(self):
         """Return whether or not parameter is available on hardware."""
-        return bool(_get_param(self.cam.handle, self.param_id, ATTR_AVAIL).value)
+        return bool(
+            _get_param(self.cam.handle, self.param_id, ATTR_AVAIL).value
+        )
 
     @property
     def count(self):
@@ -1127,12 +1286,13 @@ class PVParam():
 
 class PVEnumParam(PVParam):
     """PVParam subclass for enums"""
+
     def set_value(self, new_value):
         """Set an enum parameter value."""
         # We may be passed a value, a description string, or a tuple of
         # (value, string).
         values, descriptions = list(zip(*self.values.items()))
-        if hasattr(new_value, '__iter__'):
+        if hasattr(new_value, "__iter__"):
             desc = str(new_value[1])
         elif isinstance(new_value, str):
             desc = str(new_value)
@@ -1144,7 +1304,10 @@ class PVEnumParam(PVParam):
             new_index = descriptions.index(desc)
             new_value = values[new_index]
         elif desc:
-            raise Exception("Could not find description '%s' for enum %s." % (desc, self.name))
+            raise Exception(
+                "Could not find description '%s' for enum %s."
+                % (desc, self.name)
+            )
         super().set_value(new_value)
 
     @property
@@ -1159,13 +1322,16 @@ class PVEnumParam(PVParam):
         values = {}
         for i in range(self.count):
             length = _enum_str_length(self.cam.handle, self.param_id, i)
-            value, desc = _get_enum_param(self.cam.handle, self.param_id, i, length)
+            value, desc = _get_enum_param(
+                self.cam.handle, self.param_id, i, length
+            )
             values[value.value] = desc.value.decode()
         return values
 
 
 class PVStringParam(PVParam):
     """PVParam subclass for strings"""
+
     def set_value(self, new_value):
         """Set a string parameter value"""
         if hasattr(new_value, "encode"):
@@ -1184,11 +1350,11 @@ class PVStringParam(PVParam):
         return values
 
 
-class PVCamera(devices.FloatingDeviceMixin, devices.CameraDevice):
+class PVCamera(microscope.abc.FloatingDeviceMixin, microscope.abc.Camera):
     """Implements the CameraDevice interface for the pvcam library."""
+
     # Keep track of open cameras.
     open_cameras = []
-
 
     def __init__(self, index=0, **kwargs):
         super().__init__(index=index, **kwargs)
@@ -1201,9 +1367,9 @@ class PVCamera(devices.FloatingDeviceMixin, devices.CameraDevice):
         # Region of interest.
         self.roi = (None, None, None, None)
         # Binning setting.
-        self.binning = Binning(1, 1)
+        self.binning = microscope.Binning(1, 1)
         self._trigger = TRIG_STROBED
-        self.exposure_time = 0.001 # in seconds
+        self.exposure_time = 0.001  # in seconds
         # Cycle time
         self.cycle_time = self.exposure_time
         # Data buffer.
@@ -1214,35 +1380,46 @@ class PVCamera(devices.FloatingDeviceMixin, devices.CameraDevice):
         self._circ_buffer_length = 10
 
         # Add common settings.
-        self.add_setting('exposure time',
-                         'float',
-                         lambda: self.exposure_time,
-                         self.set_exposure_time,
-                         lambda: (1e-6, 1) )
-        self.add_setting('trigger mode',
-                         'enum',
-                         lambda: self._trigger,
-                         lambda value: setattr(self, '_trigger', value),
-                         {k: v.label for k,v in TRIGGER_MODES.items()} )
-        self.add_setting('circular buffer length',
-                         'int',
-                         lambda: self._circ_buffer_length,
-                         lambda value: setattr(self, '_circ_buffer_length', value),
-                         (2, 100))
+        self.add_setting(
+            "exposure time",
+            "float",
+            lambda: self.exposure_time,
+            self.set_exposure_time,
+            lambda: (1e-6, 1),
+        )
+        self.add_setting(
+            "trigger mode",
+            "enum",
+            lambda: self._trigger,
+            lambda value: setattr(self, "_trigger", value),
+            {k: v.label for k, v in TRIGGER_MODES.items()},
+        )
+        self.add_setting(
+            "circular buffer length",
+            "int",
+            lambda: self._circ_buffer_length,
+            lambda value: setattr(self, "_circ_buffer_length", value),
+            (2, 100),
+        )
 
     @property
     def _region(self):
         """Return a rgn_type for current roi and binning settings."""
-        return rgn_type(self.roi.left, self.roi.left + self.roi.width - 1, self.binning.h,
-                        self.roi.top, self.roi.top + self.roi.height - 1, self.binning.v)
-
+        return rgn_type(
+            self.roi.left,
+            self.roi.left + self.roi.width - 1,
+            self.binning.h,
+            self.roi.top,
+            self.roi.top + self.roi.height - 1,
+            self.binning.v,
+        )
 
     """Private methods, called here and within super classes."""
+
     def _fetch_data(self):
         """Fetch data - for use in fetch_loop."""
         # Not used: images fetched using callback.
         return None
-
 
     def _on_enable(self):
         """Enable the camera hardware and make ready to respond to triggers.
@@ -1261,6 +1438,7 @@ class PVCamera(devices.FloatingDeviceMixin, devices.CameraDevice):
             # Software triggering for single frames.
             # Set up callback.
             self._using_callback = True
+
             def cb():
                 """Soft trigger mode end-of-frame callback."""
                 timestamp = time.time()
@@ -1269,51 +1447,84 @@ class PVCamera(devices.FloatingDeviceMixin, devices.CameraDevice):
                 _exp_finish_seq(self.handle, CCS_CLEAR)
                 self._put(frame, timestamp)
                 return
+
             # Need to keep a reference to the callback.
             self._eof_callback = CALLBACK(cb)
-            _cam_register_callback(self.handle, PL_CALLBACK_EOF, self._eof_callback)
-            nbytes = _exp_setup_seq(self.handle, 1, 1, # cam, num epxosures, num regions
-                                    self._region, TRIGGER_MODES[self._trigger].pv_mode, t_exp)
-            buffer_shape = (self.roi.height//self.binning.v,
-                            self.roi.width // self.binning.h)
-            self._buffer = np.require(np.zeros(buffer_shape, dtype='uint16'),
-                                      requirements=['C_CONTIGUOUS','ALIGNED','OWNDATA'])
+            _cam_register_callback(
+                self.handle, PL_CALLBACK_EOF, self._eof_callback
+            )
+            nbytes = _exp_setup_seq(
+                self.handle,
+                1,
+                1,  # cam, num epxosures, num regions
+                self._region,
+                TRIGGER_MODES[self._trigger].pv_mode,
+                t_exp,
+            )
+            buffer_shape = (
+                self.roi.height // self.binning.v,
+                self.roi.width // self.binning.h,
+            )
+            self._buffer = np.require(
+                np.zeros(buffer_shape, dtype="uint16"),
+                requirements=["C_CONTIGUOUS", "ALIGNED", "OWNDATA"],
+            )
         else:
             # Use a circular buffer.
             self._using_callback = True
+
             def cb():
                 """Circular buffer mode end-of-frame callback."""
                 timestamp = time.time()
-                frame_p = ctypes.cast(_exp_get_latest_frame(self.handle), ctypes.POINTER(uns16))
-                frame = np.ctypeslib.as_array(frame_p, (self.roi[2], self.roi[3])).copy()
+                frame_p = ctypes.cast(
+                    _exp_get_latest_frame(self.handle), ctypes.POINTER(uns16)
+                )
+                frame = np.ctypeslib.as_array(
+                    frame_p, (self.roi[2], self.roi[3])
+                ).copy()
                 _logger.debug("Fetched frame from circular buffer.")
                 self._put(frame, timestamp)
                 return
+
             # Need to keep a reference to the callback.
             self._eof_callback = CALLBACK(cb)
-            _cam_register_callback(self.handle, PL_CALLBACK_EOF, self._eof_callback)
-            buffer_shape = (self._circ_buffer_length,
-                            self.roi.height//self.binning.v,
-                            self.roi.width // self.binning.h)
-            self._buffer = np.require(np.zeros(buffer_shape, dtype='uint16'),
-                                          requirements=['C_CONTIGUOUS', 'ALIGNED', 'OWNDATA'])
-            nbytes = _exp_setup_cont(self.handle, 1, self._region,
-                                     TRIGGER_MODES[self._trigger].pv_mode, t_exp, CIRC_OVERWRITE).value
-
-
+            _cam_register_callback(
+                self.handle, PL_CALLBACK_EOF, self._eof_callback
+            )
+            buffer_shape = (
+                self._circ_buffer_length,
+                self.roi.height // self.binning.v,
+                self.roi.width // self.binning.h,
+            )
+            self._buffer = np.require(
+                np.zeros(buffer_shape, dtype="uint16"),
+                requirements=["C_CONTIGUOUS", "ALIGNED", "OWNDATA"],
+            )
+            nbytes = _exp_setup_cont(
+                self.handle,
+                1,
+                self._region,
+                TRIGGER_MODES[self._trigger].pv_mode,
+                t_exp,
+                CIRC_OVERWRITE,
+            ).value
 
         # Read back exposure time.
         t_readback = self._params[PARAM_EXPOSURE_TIME].current
         t_resolution = self._params[PARAM_EXP_RES].current
-        multipliers = {EXP_RES_ONE_SEC: 1.,
-                       EXP_RES_ONE_MILLISEC: 1e-3,
-                       EXP_RES_ONE_MICROSEC: 1e-6}
+        multipliers = {
+            EXP_RES_ONE_SEC: 1.0,
+            EXP_RES_ONE_MILLISEC: 1e-3,
+            EXP_RES_ONE_MICROSEC: 1e-6,
+        }
         if isinstance(t_resolution, tuple):
             self.exposure_time = t_readback * multipliers[t_resolution[0]]
         else:
             self.exposure_time = t_readback * multipliers[t_resolution]
         # Update cycle time. Exposure time in seconds; readout time in microseconds.
-        self.cycle_time = self.exposure_time + 1e-6 * self._params[PARAM_READOUT_TIME].current
+        self.cycle_time = (
+            self.exposure_time + 1e-6 * self._params[PARAM_READOUT_TIME].current
+        )
         # Set up exposure time for VARIABLE_TIMED_MODE, as according to documentation.
         # It doesn't seem to work.
         if self._trigger == TRIG_VARIABLE:
@@ -1321,17 +1532,19 @@ class PVCamera(devices.FloatingDeviceMixin, devices.CameraDevice):
         # If using real triggering, start triggered acquisition.
         # (Software triggering will start acquisition in soft_trigger().)
         if self._trigger != TRIG_SOFT:
-            _exp_start_cont(self.handle, self._buffer.ctypes.data_as(ctypes.c_void_p), self._buffer.nbytes)
+            _exp_start_cont(
+                self.handle,
+                self._buffer.ctypes.data_as(ctypes.c_void_p),
+                self._buffer.nbytes,
+            )
         # Done.
         self._acquiring = True
         return self._acquiring
-
 
     def _on_disable(self):
         """Disable the hardware for a short period of inactivity."""
         self.abort()
         _cam_deregister_callback(self.handle, PL_CALLBACK_EOF)
-
 
     def _on_shutdown(self):
         """Disable the hardware for a prolonged period of inactivity."""
@@ -1342,10 +1555,10 @@ class PVCamera(devices.FloatingDeviceMixin, devices.CameraDevice):
             _logger.info("No more open cameras - calling _pvcam_uninit.")
             _pvcam_uninit()
 
-
     """Private shape-related methods. These methods do not need to account
     for camera orientation or transforms due to readout mode, as that
     is handled in the parent class."""
+
     def _get_sensor_shape(self):
         """Return the sensor shape (width, height)."""
         return self.shape
@@ -1354,31 +1567,30 @@ class PVCamera(devices.FloatingDeviceMixin, devices.CameraDevice):
         """Return the current binning (horizontal, vertical)."""
         return self.binning
 
-    @keep_acquiring
+    @microscope.abc.keep_acquiring
     def _set_binning(self, binning):
         """Set binning to (h, v)."""
         #  The keep_acquiring decorator will cause recreation of buffers.
-        self.binning = Binning(binning.h, binning.v)
+        self.binning = microscope.Binning(binning.h, binning.v)
 
     def _get_roi(self):
         """Return the current ROI (left, top, width, height)."""
         return self.roi
 
-    @keep_acquiring
+    @microscope.abc.keep_acquiring
     def _set_roi(self, roi):
         """Set the ROI to (left, tip, width, height)."""
         right = roi.left + roi.width
         bottom = roi.top + roi.height
         if (right, bottom) > self.shape:
-            raise Exception("ROI exceeds sensor area.")
+            raise ValueError("ROI exceeds sensor area.")
         self.roi = roi
 
-
     """Public methods, callable from client."""
+
     def get_id(self):
         """Get hardware's unique identifier."""
         return self._params[PARAM_HEAD_SER_NUM_ALPHA].current
-
 
     def abort(self):
         """Abort acquisition.
@@ -1392,7 +1604,6 @@ class PVCamera(devices.FloatingDeviceMixin, devices.CameraDevice):
         _exp_abort(self.handle, CCS_HALT)
         self._acquiring = False
 
-
     def initialize(self):
         """Initialise the camera."""
         # Init the DLL if necessary.
@@ -1404,11 +1615,11 @@ class PVCamera(devices.FloatingDeviceMixin, devices.CameraDevice):
         # If no cameras detected, need to deinit DLL so it can be reinited to update count.
         if _cam_get_total().value == 0:
             _pvcam_uninit()
-            raise Exception ('No cameras detected.')
+            raise microscope.InitialiseError("No cameras detected.")
         # Connect to the camera.
         _logger.info("DLL version: %s", _pvcam_get_ver().value)
         self._pv_name = _cam_get_name(self._index).value
-        _logger.info('Initializing %s', self._pv_name)
+        _logger.info("Initializing %s", self._pv_name)
         self.handle = _cam_open(self._pv_name, OPEN_EXCLUSIVE)
         PVCamera.open_cameras.append(self.handle)
         # Set up event callbacks. Tried to use the resume callback to reinit camera
@@ -1416,16 +1627,25 @@ class PVCamera(devices.FloatingDeviceMixin, devices.CameraDevice):
         # DLL throws a Windows Error 0xE06D7363.
         def _cb(event):
             _logger.info("Received %s event.", event)
-            if event == 'removed':
+            if event == "removed":
                 _logger.critical("Can not re-init hardware. Exiting.")
                 exit(-1)
             return
-        self._cbs = {'check': CALLBACK(lambda: _cb('check')),
-                     'resumed': CALLBACK(lambda: _cb('resumed')),
-                     'removed': CALLBACK(lambda: _cb('removed'))}
-        _cam_register_callback(self.handle, PL_CALLBACK_CHECK_CAMS, self._cbs['check'])
-        _cam_register_callback(self.handle, PL_CALLBACK_CAM_REMOVED, self._cbs['removed'])
-        _cam_register_callback(self.handle, PL_CALLBACK_CAM_RESUMED, self._cbs['resumed'])
+
+        self._cbs = {
+            "check": CALLBACK(lambda: _cb("check")),
+            "resumed": CALLBACK(lambda: _cb("resumed")),
+            "removed": CALLBACK(lambda: _cb("removed")),
+        }
+        _cam_register_callback(
+            self.handle, PL_CALLBACK_CHECK_CAMS, self._cbs["check"]
+        )
+        _cam_register_callback(
+            self.handle, PL_CALLBACK_CAM_REMOVED, self._cbs["removed"]
+        )
+        _cam_register_callback(
+            self.handle, PL_CALLBACK_CAM_RESUMED, self._cbs["resumed"]
+        )
         # Repopulate _params.
         self._params = {}
         for (param_id, name) in _param_to_name.items():
@@ -1446,32 +1666,43 @@ class PVCamera(devices.FloatingDeviceMixin, devices.CameraDevice):
                 raise
             except Exception as err:
                 # Test on err.args prevents indexing into empty tuple.
-                if err.args and not err.args[0].startswith('pvcam error 49'):
-                    _logger.warn('Skipping parameter %s: not supported'
-                                 ' in python.', p.name)
+                if err.args and not err.args[0].startswith("pvcam error 49"):
+                    _logger.warn(
+                        "Skipping parameter %s: not supported" " in python.",
+                        p.name,
+                    )
                     continue
-            self.add_setting(p.name,
-                             p.dtype,
-                             lambda p=p: p.current,
-                             p.set_value,
-                             lambda p=p: p.values,
-                             not p.access in [ACC_READ_WRITE, ACC_WRITE_ONLY])
+            self.add_setting(
+                p.name,
+                p.dtype,
+                lambda p=p: p.current,
+                p.set_value,
+                lambda p=p: p.values,
+                not p.access in [ACC_READ_WRITE, ACC_WRITE_ONLY],
+            )
         if PARAM_GAIN_MULT_FACTOR in self._params:
-            self.add_setting('gain',
-                             self._params[PARAM_GAIN_MULT_FACTOR].dtype,
-                             lambda: self._params[PARAM_GAIN_MULT_FACTOR].current,
-                             self._params[PARAM_GAIN_MULT_FACTOR].set_value,
-                             self._params[PARAM_GAIN_MULT_FACTOR].values)
+            self.add_setting(
+                "gain",
+                self._params[PARAM_GAIN_MULT_FACTOR].dtype,
+                lambda: self._params[PARAM_GAIN_MULT_FACTOR].current,
+                self._params[PARAM_GAIN_MULT_FACTOR].set_value,
+                self._params[PARAM_GAIN_MULT_FACTOR].values,
+            )
 
         if PARAM_PMODE in self._params:
-            self.add_setting('frame transfer mode',
-                             self._params[PARAM_PMODE].dtype,
-                             lambda: self._params[PARAM_PMODE].current,
-                             self._params[PARAM_PMODE].set_value,
-                             self._params[PARAM_PMODE].values)
+            self.add_setting(
+                "frame transfer mode",
+                self._params[PARAM_PMODE].dtype,
+                lambda: self._params[PARAM_PMODE].current,
+                self._params[PARAM_PMODE].set_value,
+                self._params[PARAM_PMODE].values,
+            )
 
-        self.shape = (self._params[PARAM_PAR_SIZE].current, self._params[PARAM_SER_SIZE].current)
-        self.roi = ROI(0, 0, self.shape[0], self.shape[1])
+        self.shape = (
+            self._params[PARAM_PAR_SIZE].current,
+            self._params[PARAM_SER_SIZE].current,
+        )
+        self.roi = microscope.ROI(0, 0, self.shape[0], self.shape[1])
 
         # Populate readout modes by iterating over readout ports and speed
         # table entries.
@@ -1481,42 +1712,47 @@ class PVCamera(devices.FloatingDeviceMixin, devices.CameraDevice):
         for i, port in ro_ports.items():
             self._params[PARAM_READOUT_PORT].set_value(i)
             ro_speeds = self._params[PARAM_SPDTAB_INDEX].values
-            for j in range(ro_speeds[0], ro_speeds[1]+1):
+            for j in range(ro_speeds[0], ro_speeds[1] + 1):
                 self._params[PARAM_SPDTAB_INDEX].set_value(j)
                 bit_depth = self._params[PARAM_BIT_DEPTH].current
                 freq = 1e9 / self._params[PARAM_PIX_TIME].current
                 if freq > 1e6:
                     freq *= 1e-6
-                    prefix = 'M'
+                    prefix = "M"
                 elif freq > 1e3:
                     freq *= 1e-3
-                    prefix = 'k'
+                    prefix = "k"
                 else:
-                    prefix = 'Hz'
-                mode_str = "%s, %s-bit, %s %sHz" % (port, bit_depth, freq, prefix)
+                    prefix = "Hz"
+                mode_str = "%s, %s-bit, %s %sHz" % (
+                    port,
+                    bit_depth,
+                    freq,
+                    prefix,
+                )
                 self._readout_modes.append(mode_str)
-                self._readout_mode_parameters.append({'port':i, 'spdtab_index': j})
+                self._readout_mode_parameters.append(
+                    {"port": i, "spdtab_index": j}
+                )
         # Set to default mode.
         self.set_readout_mode(0)
         self._params[PARAM_CLEAR_MODE].set_value(CLEAR_PRE_EXPOSURE_POST_SEQ)
 
-
-    @keep_acquiring
+    @microscope.abc.keep_acquiring
     def set_readout_mode(self, index):
         """Set the readout mode and transform."""
         params = self._readout_mode_parameters[index]
-        self._params[PARAM_READOUT_PORT].set_value(params['port'])
-        self._params[PARAM_SPDTAB_INDEX].set_value(params['spdtab_index'])
+        self._params[PARAM_READOUT_PORT].set_value(params["port"])
+        self._params[PARAM_SPDTAB_INDEX].set_value(params["spdtab_index"])
         self._readout_mode = index
         # Update transforms, if available.
         chip = self._params[PARAM_CHIP_NAME].current
         new_readout_transform = None
         readout_map = READOUT_TRANSFORMS.get(chip, None)
         if readout_map:
-            new_readout_transform = readout_map.get(params['port'], None)
+            new_readout_transform = readout_map.get(params["port"], None)
         if new_readout_transform:
             self._set_readout_transform(new_readout_transform)
-
 
     def make_safe(self):
         """Put the camera into a safe state.
@@ -1526,12 +1762,10 @@ class PVCamera(devices.FloatingDeviceMixin, devices.CameraDevice):
         if self._acquiring:
             self.abort()
 
-
-    @keep_acquiring
+    @microscope.abc.keep_acquiring
     def set_exposure_time(self, value):
         """Set the exposure time to value."""
         self.exposure_time = value
-
 
     def get_exposure_time(self):
         """Return the current exposure time.
@@ -1540,7 +1774,6 @@ class PVCamera(devices.FloatingDeviceMixin, devices.CameraDevice):
         value during _on_enable."""
         return self.exposure_time
 
-
     def get_cycle_time(self):
         """Return the cycle time.
 
@@ -1548,11 +1781,9 @@ class PVCamera(devices.FloatingDeviceMixin, devices.CameraDevice):
         value during _on_enable."""
         return self.cycle_time
 
-
     def get_trigger_type(self):
         """Return the current trigger type."""
         return TRIGGER_MODES[self._trigger].microscope_mode
-
 
     @Pyro4.oneway
     def soft_trigger(self):
@@ -1562,15 +1793,21 @@ class PVCamera(devices.FloatingDeviceMixin, devices.CameraDevice):
         Log some debugging stats in other trigger modes."""
         if self._trigger == TRIG_SOFT:
             _logger.debug("Received soft trigger ...")
-            _exp_start_seq(self.handle, self._buffer.ctypes.data_as(ctypes.c_void_p))
+            _exp_start_seq(
+                self.handle, self._buffer.ctypes.data_as(ctypes.c_void_p)
+            )
         else:
             cstatus, cbytes, cframes = _exp_check_cont_status(self.handle)
             status, bytes = _exp_check_status(self.handle)
 
-            _logger.debug('Status checks\n'
-                          'check_cont:   %s \t bytes: %d\tframes: %d\n'
-                          'check_status: %s \t bytes: %d\t',
-                          STATUS_STRINGS[cstatus.value], cbytes.value,
-                          cframes.value, STATUS_STRINGS[status.value],
-                          bytes.value)
+            _logger.debug(
+                "Status checks\n"
+                "check_cont:   %s \t bytes: %d\tframes: %d\n"
+                "check_status: %s \t bytes: %d\t",
+                STATUS_STRINGS[cstatus.value],
+                cbytes.value,
+                cframes.value,
+                STATUS_STRINGS[status.value],
+                bytes.value,
+            )
         return
