@@ -1,9 +1,8 @@
-#!/usr/bin/env python
-# -*- coding: utf-8 -*-
+#!/usr/bin/env python3
 
-## Copyright (C) 2016-2017 Mick Phillips <mick.phillips@gmail.com>
-## Copyright (C) 2017 Ian Dobbie <ian.dobbie@bioch.ox.ac.uk>
-## Copyright (C) 2019 David Miguel Susano Pinto <david.pinto@bioch.ox.ac.uk>
+## Copyright (C) 2020 David Miguel Susano Pinto <carandraug@gmail.com>
+## Copyright (C) 2020 Ian Dobbie <ian.dobbie@bioch.ox.ac.uk>
+## Copyright (C) 2020 Mick Phillips <mick.phillips@gmail.com>
 ##
 ## This file is part of Microscope.
 ##
@@ -38,6 +37,15 @@ Modifying the following settings require acquisition to be stopped:
 
 For more details, see the [XiAPI manual](https://www.ximea.com/support/wiki/apis/XiAPI_Manual#Flushing-the-queue).
 
+Hardware trigger
+----------------
+
+Ximea cameras in the MQ family accept software triggers even if set
+for hardware triggers (see `vendor issues
+#3`<https://github.com/python-microscope/vendor-issues/issues/3>).
+However, `XimeaCamera.trigger()` checks the trigger type and will
+raise an exception unless the camera is set for software triggers.
+
 """
 
 import contextlib
@@ -48,10 +56,21 @@ import typing
 import numpy as np
 from ximea import xiapi
 
-from microscope import devices
-
+import microscope
+import microscope.abc
 
 _logger = logging.getLogger(__name__)
+
+
+# The ximea package does not provide an enum for the error codes.
+# There is ximea.xidefs.ERROR_CODES which maps the error code to an
+# error message but what we need is a symbol that maps to the error
+# code so we can use while handling exceptions.
+_XI_TIMEOUT = 10
+_XI_NOT_SUPPORTED = 12
+_XI_ACQUISITION_STOPED = 45
+_XI_UNKNOWN_PARAM = 100
+
 
 # During acquisition, we rely on catching timeout errors which then
 # get discarded.  However, with debug level set to warning (XiApi
@@ -59,7 +78,7 @@ _logger = logging.getLogger(__name__)
 # making logging impossible.  So change this to error.
 #
 # Debug level is a xiapi global setting but we need a Camera instance.
-xiapi.Camera().set_debug_level('XI_DL_ERROR')
+xiapi.Camera().set_debug_level("XI_DL_ERROR")
 
 
 @contextlib.contextmanager
@@ -73,6 +92,7 @@ def _disabled_camera(camera):
             camera.enable()
     else:
         yield camera
+
 
 @contextlib.contextmanager
 def _enabled_camera(camera):
@@ -92,9 +112,9 @@ class TrgSourceMap(enum.Enum):
     # The complete list is the XI_TRG_SOURCE enum (C code) or in the
     # xidefs module (Python code).
 
-    XI_TRG_SOFTWARE = devices.TriggerType.SOFTWARE
-    XI_TRG_EDGE_RISING = devices.TriggerType.RISING_EDGE
-    XI_TRG_EDGE_FALLING = devices.TriggerType.FALLING_EDGE
+    XI_TRG_SOFTWARE = microscope.TriggerType.SOFTWARE
+    XI_TRG_EDGE_RISING = microscope.TriggerType.RISING_EDGE
+    XI_TRG_EDGE_FALLING = microscope.TriggerType.FALLING_EDGE
 
     # Not all XI_TRG_SOURCE values are defined:
     #
@@ -114,7 +134,7 @@ class TrgSelectorMap(enum.Enum):
     # xidefs module (Python code).
 
     # Trigger starts the capture of one frame.
-    XI_TRG_SEL_FRAME_START = devices.TriggerMode.ONCE
+    XI_TRG_SEL_FRAME_START = microscope.TriggerMode.ONCE
 
     # There are other modes/selector which look like they have matches
     # on TriggerMode but we never got to test them:
@@ -143,7 +163,7 @@ class TrgSelectorMap(enum.Enum):
     #   first frame.
 
 
-class XimeaCamera(devices.CameraDevice):
+class XimeaCamera(microscope.abc.TriggerTargetMixin, microscope.abc.Camera):
     """Ximea cameras
 
     Args:
@@ -151,30 +171,37 @@ class XimeaCamera(devices.CameraDevice):
             connect to.  It can be set to `None` if there is only
             camera on the system.
     """
-    def __init__(self, serial_number: typing.Optional[str] = None,
-                 **kwargs) -> None:
+
+    def __init__(
+        self, serial_number: typing.Optional[str] = None, **kwargs
+    ) -> None:
         super().__init__(**kwargs)
         self._acquiring = False
         self._handle = xiapi.Camera()
         self._img = xiapi.Image()
         self._serial_number = serial_number
         self._sensor_shape = (0, 0)
-        self._roi = devices.ROI(None,None,None,None)
-        self._binning = devices.Binning(1, 1)
+        self._roi = microscope.ROI(None, None, None, None)
+        self._binning = microscope.Binning(1, 1)
 
         # When using the Settings system, enums are not really enums
         # and even when using lists we get indices sent back and forth
         # (works fine only when using EnumInt.  The gymnastic here
         # makes it work with the rest of enums which are there to make
-        # it work with TriggerTypeMixIn.
+        # it work with TriggerTargetMixin.
         trg_source_names = [x.name for x in TrgSourceMap]
+
         def _trigger_source_setter(index: int) -> None:
             trigger_mode = TrgSourceMap[trg_source_names[index]].value
             self.set_trigger(trigger_mode, self.trigger_mode)
-        self.add_setting('trigger source', 'enum',
-                         lambda: TrgSourceMap(self.trigger_type).name,
-                         _trigger_source_setter,
-                         trg_source_names)
+
+        self.add_setting(
+            "trigger source",
+            "enum",
+            lambda: TrgSourceMap(self.trigger_type).name,
+            _trigger_source_setter,
+            trg_source_names,
+        )
 
     def _fetch_data(self) -> typing.Optional[np.ndarray]:
         if not self._acquiring:
@@ -185,23 +212,26 @@ class XimeaCamera(devices.CameraDevice):
         except Exception as err:
             # err.status may not exist so use getattr (see
             # https://github.com/python-microscope/vendor-issues/issues/2)
-            if getattr(err, 'status', None) == 10: # Timeout
+            if getattr(err, "status", None) == _XI_TIMEOUT:
                 return None
-            elif (getattr(err, 'status', None) == 45 # Acquisition is stopped
-                  and not self._acquiring):
+            elif (
+                getattr(err, "status", None) == _XI_ACQUISITION_STOPED
+                and not self._acquiring
+            ):
                 # We can end up here during disable if self._acquiring
                 # was True but is now False.
                 return None
             else:
-                raise err
+                raise
 
-        data = self._img.get_image_data_numpy() # type: np.ndarray
-        _logger.info("Fetched imaged with dims %s and size %s.",
-                     data.shape, data.size)
+        data: np.ndarray = self._img.get_image_data_numpy()
+        _logger.info(
+            "Fetched imaged with dims %s and size %s.", data.shape, data.size
+        )
         return data
 
     def abort(self):
-        _logger.info('Disabling acquisition.')
+        _logger.info("Disabling acquisition.")
         if self._acquiring:
             # We set acquiring before calling stop_acquisition because
             # the fetch loop is still running and will raise errors 45
@@ -209,7 +239,7 @@ class XimeaCamera(devices.CameraDevice):
             self._acquiring = False
             try:
                 self._handle.stop_acquisition()
-            except:
+            except Exception:
                 self._acquiring = True
                 raise
 
@@ -222,70 +252,73 @@ class XimeaCamera(devices.CameraDevice):
 
         if self._serial_number is None:
             if n_cameras > 1:
-                raise Exception('more than one Ximea camera found but the'
-                                ' serial_number argument was not specified')
-            _logger.info('serial_number is not specified but there is only one'
-                         ' camera on the system')
+                raise TypeError(
+                    "more than one Ximea camera found but the"
+                    " serial_number argument was not specified"
+                )
+            _logger.info(
+                "serial_number is not specified but there is only one"
+                " camera on the system"
+            )
             self._handle.open_device()
         else:
-            _logger.info('opening camera with serial number \'%s\'',
-                         self._serial_number)
+            _logger.info(
+                "opening camera with serial number '%s'", self._serial_number
+            )
             self._handle.open_device_by_SN(self._serial_number)
-            # Camera.dev_id defaults to zero.  However, after opening
-            # the device by serial number is is not updated (see
-            # https://github.com/python-microscope/vendor-issues/issues/1).
-            # So we manually iterate over each possible device ID and
-            # modify dev_id until it behaves as it should.  If we
-            # don't fix this and there are multiple cameras connected,
-            # some of the handle methods will return info from another
-            # camera.
-            for dev_id in range(n_cameras):
-                self._handle.dev_id = dev_id
-                if (self._serial_number.encode()
-                    == self._handle.get_device_info_string('device_sn')):
-                    break
-            else:
-                raise Exception('failed to get DevId for device with SN %s'
-                                % self._serial_number)
 
-        self._sensor_shape = (self._handle.get_width_maximum()
-                              + self._handle.get_offsetX_maximum(),
-                              self._handle.get_height_maximum()
-                              + self._handle.get_offsetY_maximum())
-        self._roi = devices.ROI(left=0, top=0,
-                                width=self._sensor_shape[0],
-                                height=self._sensor_shape[1])
+        self._sensor_shape = (
+            self._handle.get_width_maximum()
+            + self._handle.get_offsetX_maximum(),
+            self._handle.get_height_maximum()
+            + self._handle.get_offsetY_maximum(),
+        )
+        self._roi = microscope.ROI(
+            left=0,
+            top=0,
+            width=self._sensor_shape[0],
+            height=self._sensor_shape[1],
+        )
         self.set_roi(self._roi)
 
-        self.set_trigger(devices.TriggerType.SOFTWARE,
-                         devices.TriggerMode.ONCE)
+        self.set_trigger(
+            microscope.TriggerType.SOFTWARE, microscope.TriggerMode.ONCE
+        )
 
         # When we return the sensor temperature we want to return the
         # temperature that's closest to the chip since that's the one
         # that has the biggest impact on image noise.  We don't know
         # what temperature sensors each camera has so we try one at a
         # time, by order of preference, until it works.
-        for temperature_selector in ('XI_TEMP_IMAGE_SENSOR_DIE',
-                                     'XI_TEMP_IMAGE_SENSOR_DIE_RAW',
-                                     'XI_TEMP_SENSOR_BOARD',
-                                     'XI_TEMP_INTERFACE_BOARD',
-                                     'XI_TEMP_FRONT_HOUSING',
-                                     'XI_TEMP_REAR_HOUSING',
-                                     'XI_TEMP_TEC1_COLD',
-                                     'XI_TEMP_TEC1_HOT'):
+        for temperature_selector in (
+            "XI_TEMP_IMAGE_SENSOR_DIE",
+            "XI_TEMP_IMAGE_SENSOR_DIE_RAW",
+            "XI_TEMP_SENSOR_BOARD",
+            "XI_TEMP_INTERFACE_BOARD",
+            "XI_TEMP_FRONT_HOUSING",
+            "XI_TEMP_REAR_HOUSING",
+            "XI_TEMP_TEC1_COLD",
+            "XI_TEMP_TEC1_HOT",
+        ):
             try:
                 self._handle.set_temp_selector(temperature_selector)
             except xiapi.Xi_error as err:
-                if getattr(err, 'status', None) == 12: # Not supported
-                    _logger.info('no hardware support for %s temperature'
-                                 ' readings', temperature_selector)
+                # We need to catch both "not supported" and "unknown
+                # parameter" but we don't understand their difference.
+                # We can definitely get both (see issue #169).
+                status = getattr(err, "status", None)
+                if status in [_XI_NOT_SUPPORTED, _XI_UNKNOWN_PARAM]:
+                    _logger.info(
+                        "no hardware support for %s temperature" " readings",
+                        temperature_selector,
+                    )
                 else:
                     raise
             else:
-                _logger.info('temperature reading set to %s',
-                             temperature_selector)
+                _logger.info(
+                    "temperature reading set to %s", temperature_selector
+                )
                 break
-
 
     def make_safe(self):
         if self._acquiring:
@@ -307,44 +340,35 @@ class XimeaCamera(devices.CameraDevice):
     def set_exposure_time(self, value: float) -> None:
         # exposure times are set in us.
         try:
-            self._handle.set_exposure(int(value * 1000000))
+            self._handle.set_exposure_direct(int(value * 1000000))
         except Exception as err:
             _logger.debug("set_exposure_time exception: %s", err)
 
     def get_exposure_time(self) -> float:
         # exposure times are in us, so multiple by 1E-6 to get seconds.
-        return (self._handle.get_exposure() * 1.0E-6)
+        return self._handle.get_exposure() * 1.0e-6
 
     def get_cycle_time(self):
-        return (1.0/self._handle.get_framerate())
+        return 1.0 / self._handle.get_framerate()
 
     def _get_sensor_shape(self) -> typing.Tuple[int, int]:
         return self._sensor_shape
 
-
     def get_sensor_temperature(self) -> float:
         return self._handle.get_temp()
 
-
     def soft_trigger(self) -> None:
-        # Cameras in the MQ family have the "feature" of accepting
-        # software triggers even when set for hardware triggering.
-        # Because of this, so we need to check this ourselves.  See
+        self.trigger()
+
+    def _do_trigger(self) -> None:
+        # Value for set_trigger_software() has no meaning.  See
         # https://github.com/python-microscope/vendor-issues/issues/3
-        if self.trigger_type is not devices.TriggerType.SOFTWARE:
-            _logger.warning('ignoring soft_trigger() because camera is not'
-                            ' in software trigger type')
-        else:
-            _logger.debug('sending software trigger')
-            # Value for set_trigger_software() has no meaning.  See
-            # https://github.com/python-microscope/vendor-issues/issues/3
-            self._handle.set_trigger_software(1)
+        self._handle.set_trigger_software(1)
 
-
-    def _get_binning(self) -> devices.Binning:
+    def _get_binning(self) -> microscope.Binning:
         return self._binning
 
-    def _set_binning(self, binning: devices.Binning) -> bool:
+    def _set_binning(self, binning: microscope.Binning) -> bool:
         if binning == self._binning:
             return True
         # We don't have a ximea camera that supports binning so we
@@ -353,20 +377,24 @@ class XimeaCamera(devices.CameraDevice):
         # what XiAPI does when mixing ROI and binning.
         raise NotImplementedError()
 
-
-    def _get_roi(self) -> devices.ROI:
-        assert self._roi == devices.ROI(self._handle.get_offsetX(),
-                                        self._handle.get_offsetY(),
-                                        self._handle.get_width(),
-                                        self._handle.get_height()), \
-            "ROI attribute is out of sync with internal camera setting"
+    def _get_roi(self) -> microscope.ROI:
+        assert self._roi == microscope.ROI(
+            self._handle.get_offsetX(),
+            self._handle.get_offsetY(),
+            self._handle.get_width(),
+            self._handle.get_height(),
+        ), "ROI attribute is out of sync with internal camera setting"
         return self._roi
 
-    def _set_roi(self, roi: devices.ROI) -> bool:
-        if (roi.width + roi.left > self._sensor_shape[0]
-            or roi.height + roi.top > self._sensor_shape[1]):
-            raise ValueError('ROI %s does not fit in sensor shape %s'
-                             % (roi, self._sensor_shape))
+    def _set_roi(self, roi: microscope.ROI) -> bool:
+        if (
+            roi.width + roi.left > self._sensor_shape[0]
+            or roi.height + roi.top > self._sensor_shape[1]
+        ):
+            raise ValueError(
+                "ROI %s does not fit in sensor shape %s"
+                % (roi, self._sensor_shape)
+            )
         try:
             # These methods will fail if the width/height plus their
             # corresponding offsets are higher than the sensor size.
@@ -381,12 +409,11 @@ class XimeaCamera(devices.CameraDevice):
                 self._handle.set_height(roi.height)
                 self._handle.set_offsetX(roi.left)
                 self._handle.set_offsetY(roi.top)
-        except:
-            self._set_roi(self._roi) # set it back to whatever was before
+        except Exception:
+            self._set_roi(self._roi)  # set it back to whatever was before
             raise
         self._roi = roi
         return True
-
 
     def _on_shutdown(self) -> None:
         if self._acquiring:
@@ -398,39 +425,54 @@ class XimeaCamera(devices.CameraDevice):
             # device handler is NULL.
             self._handle.close_device()
         else:
-            _logger.warning('shutdown() called but camera was already closed')
+            _logger.warning("shutdown() called but camera was already closed")
 
     @property
-    def trigger_mode(self) -> devices.TriggerMode:
+    def trigger_mode(self) -> microscope.TriggerMode:
         trg_selector = self._handle.get_trigger_selector()
         try:
             tmode = TrgSelectorMap[trg_selector]
         except KeyError:
-            raise Exception('somehow set to unsupported trigger mode %s'
-                            % trg_selector)
+            raise Exception(
+                "somehow set to unsupported trigger mode %s" % trg_selector
+            )
         return tmode.value
 
     @property
-    def trigger_type(self) -> devices.TriggerType:
+    def trigger_type(self) -> microscope.TriggerType:
         trg_source = self._handle.get_trigger_source()
         try:
             ttype = TrgSourceMap[trg_source]
         except KeyError:
-            raise Exception('somehow set to unsupported trigger type %s'
-                            % trg_source)
+            raise Exception(
+                "somehow set to unsupported trigger type %s" % trg_source
+            )
         return ttype.value
 
-    def set_trigger(self, ttype: devices.TriggerType,
-                    tmode: devices.TriggerMode) -> None:
-        if tmode is not devices.TriggerMode.ONCE:
-            raise Exception('%s not supported (only TriggerMode.ONCE)' % tmode)
+    def set_trigger(
+        self, ttype: microscope.TriggerType, tmode: microscope.TriggerMode
+    ) -> None:
+        if tmode is not microscope.TriggerMode.ONCE:
+            raise microscope.UnsupportedFeatureError(
+                "%s not supported (only TriggerMode.ONCE)" % tmode
+            )
 
         try:
             trg_source = TrgSourceMap(ttype)
         except ValueError:
-            raise Exception('no support for trigger type %s' % ttype)
+            raise microscope.UnsupportedFeatureError(
+                "no support for trigger type %s" % ttype
+            )
 
         if trg_source.name != self._handle.get_trigger_source():
             # Changing trigger source requires stopping acquisition.
             with _disabled_camera(self):
                 self._handle.set_trigger_source(trg_source.name)
+
+    def get_trigger_type(self) -> int:
+        ttype_microscope_to_cockpit = {
+            microscope.TriggerType.SOFTWARE: microscope.abc.TRIGGER_SOFT,
+            microscope.TriggerType.RISING_EDGE: microscope.abc.TRIGGER_BEFORE,
+            microscope.TriggerType.FALLING_EDGE: microscope.abc.TRIGGER_AFTER,
+        }
+        return ttype_microscope_to_cockpit[self.trigger_type]
