@@ -1,7 +1,8 @@
-#!/usr/bin/env python
-# -*- coding: utf-8 -*-
+#!/usr/bin/env python3
 
-## Copyright (C) 2019 David Miguel Susano Pinto <david.pinto@bioch.ox.ac.uk>
+## Copyright (C) 2020 David Miguel Susano Pinto <carandraug@gmail.com>
+##
+## This file is part of Microscope.
 ##
 ## Microscope is free software: you can redistribute it and/or modify
 ## it under the terms of the GNU General Public License as published by
@@ -33,11 +34,14 @@ import typing
 
 import serial
 
-import microscope.devices
+import microscope
+import microscope.abc
 
 
+# TODO: move this into its own module to be used by others.
 class _SyncSerial:
     """Wraps a `Serial` instance with a lock for synchronization."""
+
     def __init__(self, serial: serial.Serial) -> None:
         self._serial = serial
         self._lock = threading.RLock()
@@ -49,6 +53,10 @@ class _SyncSerial:
     def readline(self) -> bytes:
         with self._lock:
             return self._serial.readline()
+
+    def readlines(self, hint: int = -1) -> typing.List[bytes]:
+        with self._lock:
+            return self._serial.readlines(hint)
 
     def write(self, data: bytes) -> int:
         with self._lock:
@@ -62,25 +70,28 @@ class _SpectraIIIConnection:
     only tested for it.  But it should work with other lumencor light
     engines with little work though, if only we got access to them.
     """
+
     def __init__(self, serial: _SyncSerial) -> None:
         self._serial = serial
         # We use command() and readline() instead of get_command() in
         # case this is not a Lumencor and won't even give a standard
         # answer and raises an exception during the answer validation.
-        self._serial.write(b'GET MODEL\n')
+        self._serial.write(b"GET MODEL\n")
         answer = self._serial.readline()
-        if not answer.startswith(b'A MODEL Spectra III'):
-            raise RuntimeError("Not a Lumencor Spectra III Light Engine")
+        if not answer.startswith(b"A MODEL Spectra III"):
+            raise microscope.InitialiseError(
+                "Not a Lumencor Spectra III Light Engine"
+            )
 
     def command_and_answer(self, *TX_tokens: bytes) -> bytes:
         # Command contains two or more tokens.  The first token for a
         # TX (transmitted) command string is one of the two keywords
         # GET, SET (to query or to set).  The second token is the
         # command name.
-        assert len(TX_tokens) >= 2, 'invalid command with less than two tokens'
-        assert TX_tokens[0] in (b'GET', b'SET'), 'invalid command (not SET/GET)'
+        assert len(TX_tokens) >= 2, "invalid command with less than two tokens"
+        assert TX_tokens[0] in (b"GET", b"SET"), "invalid command (not SET/GET)"
 
-        TX_command = b' '.join(TX_tokens) + b'\n'
+        TX_command = b" ".join(TX_tokens) + b"\n"
         with self._serial.lock:
             self._serial.write(TX_command)
             answer = self._serial.readline()
@@ -88,82 +99,70 @@ class _SpectraIIIConnection:
         # A received answer has at least two tokens.  The first token
         # is A or E (for success or failure).  The second token is the
         # command name (second token of the transmitted command).
-        if (len(RX_tokens) < 2 or RX_tokens[0] != b'A'
-            or RX_tokens[1] != TX_tokens[1]):
-            raise RuntimeError('command %s failed: %s' % (TX_command, answer))
+        if (
+            len(RX_tokens) < 2
+            or RX_tokens[0] != b"A"
+            or RX_tokens[1] != TX_tokens[1]
+        ):
+            raise microscope.DeviceError(
+                "command %s failed: %s" % (TX_command, answer)
+            )
         return answer
 
     def get_command(self, command: bytes, *args: bytes) -> bytes:
-        answer = self.command_and_answer(b'GET', command, *args)
+        answer = self.command_and_answer(b"GET", command, *args)
         # The three bytes we remove at the start are the 'A ' before
         # the command, and the space after the command.  The last two
         # bytes are '\r\n'.
-        return answer[3+len(command):-2]
+        return answer[3 + len(command) : -2]
 
     def set_command(self, command: bytes, *args: bytes) -> None:
-        self.command_and_answer(b'SET', command, *args)
+        self.command_and_answer(b"SET", command, *args)
 
     def get_channel_map(self) -> typing.List[typing.Tuple[int, str]]:
-        answer = self.get_command(b'CHMAP')
+        answer = self.get_command(b"CHMAP")
         return list(enumerate(answer.decode().split()))
 
 
 class _LightChannelConnection:
     """Commands for a channel in a Lumencor light engine."""
+
     def __init__(self, connection: _SpectraIIIConnection, index: int) -> None:
         self._conn = connection
-        self._index_bytes = b'%d' % index
+        self._index_bytes = b"%d" % index
 
     def get_light_state(self) -> bool:
         """On (True) or off (False) state"""
         # We use CHACT (actual light state) instead of CH (light
         # state) because CH checks both the TTL inputs and channel
         # state switches.
-        state = self._conn.get_command(b'CHACT', self._index_bytes)
-        if state == b'1':
+        state = self._conn.get_command(b"CHACT", self._index_bytes)
+        if state == b"1":
             return True
-        elif state == b'0':
+        elif state == b"0":
             return False
         else:
-            raise RuntimeError('unexpected answer')
+            raise microscope.DeviceError("unexpected answer")
 
     def set_light_state(self, state: bool) -> None:
         """Turn light on (True) or off (False)."""
-        state_arg = b'1' if state else b'0'
-        self._conn.set_command(b'CH', self._index_bytes, state_arg)
+        state_arg = b"1" if state else b"0"
+        self._conn.set_command(b"CH", self._index_bytes, state_arg)
 
     def get_max_intensity(self) -> int:
         """Maximum valid intensity that can be applied to a light channel."""
-        return int(self._conn.get_command(b'MAXINT', self._index_bytes))
+        return int(self._conn.get_command(b"MAXINT", self._index_bytes))
 
-    def get_power_output(self) -> float:
-        """Estimated power output for a given channel (in mW).
-
-        Power estimation is based on the power output calibration
-        factor, power sensor reading, power sensor exposure, power
-        sensor gain and crosstalk level.  Estimation model assumes
-        linear dependency.
-        """
-        return float(self._conn.get_command(b'CHPWRWATTS', self._index_bytes))
-
-    def get_power_reference(self) -> float:
-        """Light power reference (in mW).
-
-        This gives an indication of the power output when the channel
-        is set to its maximum intensity.  However, its dependent on
-        being manually set and kept adjusted over time.
-
-        A value of -1 is returned if the power reference hasn't been
-        defined yet.
-        """
-        return float(self._conn.get_command(b'PWRREF', self._index_bytes))
+    def get_intensity(self) -> int:
+        """Current intensity setting between 0 and maximum intensity."""
+        return int(self._conn.get_command(b"CHINT", self._index_bytes))
 
     def set_intensity(self, intensity: int) -> None:
         """Set light intensity between 0 and maximum intensity."""
-        self._conn.set_command(b'CHINT', self._index_bytes, b'%d' % intensity)
+        self._conn.set_command(b"CHINT", self._index_bytes, b"%d" % intensity)
 
 
-class SpectraIIILightEngine(microscope.devices.ControllerDevice):
+class SpectraIIILightEngine(microscope.abc.Controller):
     """Spectra III Light Engine.
 
     Args:
@@ -182,75 +181,63 @@ class SpectraIIILightEngine(microscope.devices.ControllerDevice):
     are already on, or by turning on additional sources, commands will
     be rejected. To clear the error condition, reduce intensities of
     sources that are on or turn off additional sources.
-
-    .. note::
-
-       This relies on having power reference values set for each
-       channel.  The Spectra light engines do not provide a method to
-       obtain the maximum power output or to set the power ouput.  As
-       such, this relies on the internal power reference value.  This
-       should be set manually on the device to obtain reasonable
-       results when setting the power output.
-
     """
+
     def __init__(self, port: str, **kwargs) -> None:
         super().__init__(**kwargs)
-        self._lights = {} # type: typing.Mapping[str, microscope.devices.Device]
+        self._lights: typing.Mapping[str, microscope.abc.Device] = {}
 
         # We use standard (not legacy) mode communication so 115200,8,N,1
-        serial_conn = serial.Serial(port=port, baudrate=115200, timeout=1,
-                                    bytesize=serial.EIGHTBITS,
-                                    stopbits=serial.STOPBITS_ONE,
-                                    parity=serial.PARITY_NONE, xonxoff=False,
-                                    rtscts=False, dsrdtr=False)
+        serial_conn = serial.Serial(
+            port=port,
+            baudrate=115200,
+            timeout=1,
+            bytesize=serial.EIGHTBITS,
+            stopbits=serial.STOPBITS_ONE,
+            parity=serial.PARITY_NONE,
+            xonxoff=False,
+            rtscts=False,
+            dsrdtr=False,
+        )
         connection = _SpectraIIIConnection(_SyncSerial(serial_conn))
 
         for index, name in connection.get_channel_map():
-            if name in self._lights:
-                raise RuntimeError('multiple lights with name \'%s\'' % name)
+            assert (
+                name not in self._lights
+            ), "light with name '%s' already mapped"
             self._lights[name] = _SpectraIIILightChannel(connection, index)
 
     @property
-    def devices(self) -> typing.Mapping[str, microscope.devices.Device]:
+    def devices(self) -> typing.Mapping[str, microscope.abc.Device]:
         return self._lights
 
 
-class _SpectraIIILightChannel(microscope.devices.LaserDevice):
+class _SpectraIIILightChannel(microscope.abc.LightSource):
     """A single light channel from a light engine.
 
-    A channel is not necessarily a lasers although it subclasses from
-    `LaserDevice`.  Constituent light sources may include LEDs,
-    luminescent light pipes, or lasers.
+    A channel may be an LED, luminescent light pipe, or a laser.
     """
+
     def __init__(self, connection: _SpectraIIIConnection, index: int) -> None:
         super().__init__()
         self._conn = _LightChannelConnection(connection, index)
         # The lumencor only allows to set the power via intensity
-        # levels (values between 0 and MAXINT) .  There is no method
-        # to query the maximum power output, that information is on
-        # the device certificate of conformance and may changes over
-        # time.
-        #
-        # Power Reference is close to the max possible power (mw).  It
-        # seems to do nothing other than providing a estimate of what
-        # power will be emitted when the intensity is at its maximum.
-        # It needs to be set manually and kept up to date.
-        self._power_ref = self._conn.get_power_reference() # type: float
-        if self._power_ref == -1:
-            raise RuntimeError('Power reference value is not set')
-        self._max_intensity = self._conn.get_max_intensity() # type: int
+        # levels (values between 0 and MAXINT).  We keep the max
+        # intensity internal as float for the multiply/divide
+        # operations.
+        self._max_intensity = float(self._conn.get_max_intensity())
 
     def initialize(self) -> None:
         pass
 
     def _on_shutdown(self) -> None:
         # There is a shutdown command but this actually powers off the
-        # device which is not what LaserDevice.shutdown() is meant to
+        # device which is not what LightSource.shutdown() is meant to
         # do.  So do nothing.
         pass
 
     def get_status(self) -> typing.List[str]:
-        status = [] # type: typing.List[str]
+        status: typing.List[str] = []
         return status
 
     def enable(self) -> None:
@@ -262,17 +249,8 @@ class _SpectraIIILightChannel(microscope.devices.LaserDevice):
     def get_is_on(self) -> bool:
         return self._conn.get_light_state()
 
-    def get_min_power_mw(self) -> float:
-        return 0.0
+    def _do_set_power(self, power: float) -> None:
+        self._conn.set_intensity(int(power * self._max_intensity))
 
-    def get_max_power_mw(self) -> float:
-        return self._power_ref
-
-    def get_power_mw(self) -> float:
-        return self._conn.get_power_output()
-
-    def _set_power_mw(self, mw: float) -> None:
-        # The mw argument should have already been clipped by
-        # `LaserDevice.set_power_mw()` so no need to do so again.
-        intensity = int((mw / self._power_ref) * self._max_intensity)
-        self._conn.set_intensity(intensity)
+    def _do_get_power(self) -> float:
+        return self._conn.get_intensity() / self._max_intensity
