@@ -1,7 +1,8 @@
-#!/usr/bin/env python
-# -*- coding: utf-8 -*-
+#!/usr/bin/env python3
 
-## Copyright (C) 2017 David Pinto <david.pinto@bioch.ox.ac.uk>
+## Copyright (C) 2020 David Miguel Susano Pinto <carandraug@gmail.com>
+##
+## This file is part of Microscope.
 ##
 ## Microscope is free software: you can redistribute it and/or modify
 ## it under the terms of the GNU General Public License as published by
@@ -25,12 +26,21 @@ import warnings
 
 import numpy
 
-from microscope.devices import DeformableMirror
+import microscope
+import microscope.abc
 
-import microscope._wrappers.BMC as BMC
+try:
+    import microscope._wrappers.BMC as BMC
+except Exception as e:
+    raise microscope.LibraryLoadError(e) from e
 
 
-class BMCDeformableMirror(DeformableMirror):
+class BMCDeformableMirror(microscope.abc.DeformableMirror):
+    """Boston MicroMachines (BMC) deformable mirror.
+
+    BMC deformable mirrors only support software trigger.
+    """
+
     def __init__(self, serial_number: str, **kwargs) -> None:
         super().__init__(**kwargs)
         self._dm = BMC.DM()
@@ -42,18 +52,37 @@ class BMCDeformableMirror(DeformableMirror):
 
         status = BMC.Open(self._dm, serial_number.encode())
         if status:
-            raise Exception(BMC.ErrorString(status))
+            raise microscope.InitialiseError(BMC.ErrorString(status))
 
     @property
     def n_actuators(self) -> int:
         return self._dm.ActCount
 
-    def apply_pattern(self, pattern: numpy.ndarray) -> None:
-        self._validate_patterns(pattern)
+    @property
+    def trigger_type(self) -> microscope.TriggerType:
+        return microscope.TriggerType.SOFTWARE
+
+    @property
+    def trigger_mode(self) -> microscope.TriggerMode:
+        return microscope.TriggerMode.ONCE
+
+    def set_trigger(
+        self, ttype: microscope.TriggerType, tmode: microscope.TriggerMode
+    ) -> None:
+        if ttype is not microscope.TriggerType.SOFTWARE:
+            raise microscope.UnsupportedFeatureError(
+                "the only trigger type supported is software"
+            )
+        if tmode is not microscope.TriggerMode.ONCE:
+            raise microscope.UnsupportedFeatureError(
+                "the only trigger mode supported is 'once'"
+            )
+
+    def _do_apply_pattern(self, pattern: numpy.ndarray) -> None:
         data_pointer = pattern.ctypes.data_as(ctypes.POINTER(ctypes.c_double))
         status = BMC.SetArray(self._dm, data_pointer, None)
         if status:
-            raise Exception(BMC.ErrorString(status))
+            raise microscope.DeviceError(BMC.ErrorString(status))
 
     def __del__(self) -> None:
         status = BMC.Close(self._dm)
