@@ -1,7 +1,6 @@
-#!/usr/bin/env python
-# -*- coding: utf-8 -*-
+#!/usr/bin/env python3
 
-## Copyright (C) 2017 David Pinto <david.pinto@bioch.ox.ac.uk>
+## Copyright (C) 2020 David Miguel Susano Pinto <carandraug@gmail.com>
 ##
 ## This file is part of Microscope.
 ##
@@ -26,13 +25,13 @@ subclasses from that device type class of tests.  Each such class only
 needs to implement the `setUp` method.  It may also add device
 specific tests.
 
-Using lasers as example, there is a :class:`.LaserTests` class full of
-`test_*` methods, each of them a test on its own.  For each laser
-device supported there is one test class, e.g.,
+Using lasers as example, there is a :class:`.LightSourceTests` class
+full of `test_*` methods, each of them a test on its own.  For each
+light source device supported there is one test class, e.g.,
 `TestOmicronDeepstarLaser`, and `TestCoherentSapphireLaser`.  These
-subclass from both :class:`unittest.TestCase` and `LaserTests` and
-need only to implement `setUp` which sets up the fake and constructs
-the device instance required to run the tests.
+subclass from both :class:`unittest.TestCase` and `LightSourceTests`
+and need only to implement `setUp` which sets up the fake and
+constructs the device instance required to run the tests.
 
 """
 
@@ -46,49 +45,51 @@ import microscope.testsuite.mock_devices as mocks
 
 
 class TestSerialMock(unittest.TestCase):
-    ## Our tests for serial devices depend on our SerialMock base
-    ## class working properly so yeah, we need tests for that too.
+    # Our tests for serial devices depend on our SerialMock base
+    # class working properly so yeah, we need tests for that too.
     class Serial(mocks.SerialMock):
-        eol = b'\r\n'
+        eol = b"\r\n"
+
         def handle(self, command):
-            if command.startswith(b'echo '):
+            if command.startswith(b"echo "):
                 self.in_buffer.write(command[5:] + self.eol)
-            elif command in [b'foo', b'bar']:
+            elif command in [b"foo", b"bar"]:
                 pass
             else:
                 raise RuntimeError("unknown command '%s'" % command.decode())
 
     def setUp(self):
         self.serial = TestSerialMock.Serial()
-        patcher = unittest.mock.patch.object(TestSerialMock.Serial, 'handle',
-                                             wraps=self.serial.handle)
+        patcher = unittest.mock.patch.object(
+            TestSerialMock.Serial, "handle", wraps=self.serial.handle
+        )
         self.addCleanup(patcher.stop)
         self.mock = patcher.start()
 
     def test_simple_commands(self):
-        self.serial.write(b'foo\r\n')
-        self.mock.assert_called_once_with(b'foo')
+        self.serial.write(b"foo\r\n")
+        self.mock.assert_called_once_with(b"foo")
 
     def test_partial_commands(self):
-        self.serial.write(b'fo')
-        self.serial.write(b'o')
-        self.serial.write(b'\r\n')
-        self.serial.handle.assert_called_once_with(b'foo')
+        self.serial.write(b"fo")
+        self.serial.write(b"o")
+        self.serial.write(b"\r\n")
+        self.serial.handle.assert_called_once_with(b"foo")
 
     def test_multiple_commands(self):
-        self.serial.write(b'foo\r\nbar\r\n')
-        calls = [unittest.mock.call(x) for x in [b'foo', b'bar']]
+        self.serial.write(b"foo\r\nbar\r\n")
+        calls = [unittest.mock.call(x) for x in [b"foo", b"bar"]]
         self.assertEqual(self.serial.handle.mock_calls, calls)
 
     def test_unix_eol(self):
-        self.serial.eol = b'\n'
-        self.serial.write(b'foo\nbar\n')
-        calls = [unittest.mock.call(x) for x in [b'foo', b'bar']]
+        self.serial.eol = b"\n"
+        self.serial.write(b"foo\nbar\n")
+        calls = [unittest.mock.call(x) for x in [b"foo", b"bar"]]
         self.assertEqual(self.serial.handle.mock_calls, calls)
 
     def test_write(self):
-        self.serial.write(b'echo qux\r\n')
-        self.assertEqual(self.serial.readline(), b'qux\r\n')
+        self.serial.write(b"echo qux\r\n")
+        self.assertEqual(self.serial.readline(), b"qux\r\n")
 
 
 class DeviceTests:
@@ -109,15 +110,15 @@ class DeviceTests:
         self.device.shutdown()
 
     def test_enable_and_disable(self):
-        ## TODO: we need to define what happens when enable is called
-        ## and device has not been initialised.  See issue #69
+        # TODO: we need to define what happens when enable is called
+        # and device has not been initialised.  See issue #69
         self.device.initialize()
         self.device.enable()
         self.assertTrue(self.device.enabled)
-        ## We don't check if it is disabled after shutdown because
-        ## some devices can't be turned off.
-        ## TODO: add a `has_disabled_state` to the fake so we can
-        ## query whether we can check about being disabled.
+        # We don't check if it is disabled after shutdown because
+        # some devices can't be turned off.
+        # TODO: add a `has_disabled_state` to the fake so we can
+        # query whether we can check about being disabled.
         self.device.disable()
         self.device.shutdown()
 
@@ -179,15 +180,15 @@ class SerialDeviceTests:
         self.assertEqual(self.device.connection.dsrdtr, self.fake.dsrdtr)
 
 
-class LaserTests(DeviceTests):
-    """Base class for :class:`LaserDevice` tests.
+class LightSourceTests(DeviceTests):
+    """Base class for :class:`LightSource` tests.
 
     This class implements all the general laser tests and is meant to
     be mixed with :class:`unittest.TestCase`.  Subclasses must
     implement the `setUp` method which must add two properties:
 
     `device`
-        Instance of the :class:`LaserDevice` implementation being
+        Instance of the :class:`LightSource` implementation being
         tested.
 
     `fake`
@@ -197,9 +198,10 @@ class LaserTests(DeviceTests):
         class that fakes the hardware.
 
     """
+
     def assertEqualMW(self, first, second, msg=None):
-        ## We could be smarter, but rounding the values should be
-        ## enough to check the values when comparing power levels.
+        # We could be smarter, but rounding the values should be
+        # enough to check the values when comparing power levels.
         self.assertEqual(round(first), round(second), msg)
 
     def test_being(self):
@@ -213,10 +215,10 @@ class LaserTests(DeviceTests):
         self.assertEqual(self.device.connection.light, self.device.get_is_on())
 
     def test_off_after_constructor(self):
-        ## Some lasers, such as our Coherent Sapphire emit laser
-        ## radiation as soon as the key is switched on.  We should
-        ## ensure that the laser is turned off during the
-        ## construction.
+        # Some lasers, such as our Coherent Sapphire emit laser
+        # radiation as soon as the key is switched on.  We should
+        # ensure that the laser is turned off during the
+        # construction.
         self.assertFalse(self.device.get_is_on())
 
     def test_turning_on_and_off(self):
@@ -230,45 +232,36 @@ class LaserTests(DeviceTests):
         self.device.disable()
         self.device.shutdown()
 
-    def test_query_power_range(self):
-        min_mw = self.device.get_min_power_mw()
-        max_mw = self.device.get_max_power_mw()
-        self.assertIsInstance(min_mw, float)
-        self.assertIsInstance(max_mw, float)
-        self.assertEqualMW(min_mw, self.fake.min_power)
-        self.assertEqualMW(max_mw, self.fake.max_power)
-
     def test_power_when_off(self):
         self.device.disable()
-        power = self.device.get_power_mw()
-        self.assertIsInstance(power, float)
-        self.assertEqual(power, 0.0)
+        self.assertIsInstance(self.device.power, float)
+        self.assertEqual(self.device.power, 0.0)
 
     def test_setting_power(self):
         self.device.enable()
-        power = self.device.get_power_mw()
-        self.assertIsInstance(power, float)
-        self.assertEqualMW(power, self.fake.default_power)
-        self.assertEqualMW(power, self.device.get_set_power_mw())
+        self.assertIsInstance(self.device.power, float)
+        power_mw = self.device.power * self.fake.max_power
+        self.assertEqualMW(power_mw, self.fake.default_power)
+        self.assertEqualMW(self.device.power, self.device.get_set_power())
 
-        new_power = (self.fake.min_power
-                     + ((self.fake.max_power - self.fake.min_power) /2.0))
-        self.device.set_power_mw(new_power)
-        self.assertEqualMW(self.device.get_power_mw(), new_power)
-        self.assertEqualMW(new_power, self.device.get_set_power_mw())
+        new_power = 0.5
+        new_power_mw = new_power * self.fake.max_power
+        self.device.power = new_power
+        self.assertEqualMW(
+            self.device.power * self.fake.max_power, new_power_mw
+        )
+        self.assertEqualMW(new_power, self.device.get_set_power())
 
     def test_setting_power_outside_limit(self):
         self.device.enable()
-        below_limit = self.fake.min_power - 10.0
-        above_limit = self.fake.max_power + 10.0
-        self.device.set_power_mw(below_limit)
-        self.assertEqualMW(self.device.get_power_mw(),
-                           self.device.get_min_power_mw(),
-                           'clip setting power to the valid range')
-        self.device.set_power_mw(above_limit)
-        self.assertEqualMW(self.device.get_power_mw(),
-                           self.device.get_max_power_mw(),
-                           'clip setting power to the valid range')
+        self.device.power = -0.1
+        self.assertEqual(
+            self.device.power,
+            self.fake.min_power / self.fake.max_power,
+            "clip setting power below 0",
+        )
+        self.device.power = 1.1
+        self.assertEqual(self.device.power, 1.0, "clip setting power above 1")
 
     def test_status(self):
         status = self.device.get_status()
@@ -281,8 +274,24 @@ class CameraTests(DeviceTests):
     pass
 
 
-class FilterWheelTests(DeviceTests):
+class ControllerTests(DeviceTests):
     pass
+
+
+class FilterWheelTests(DeviceTests):
+    def test_get_and_set_position(self):
+        self.assertEqual(self.device.position, 0)
+        max_pos = self.device.n_positions - 1
+        self.device.position = max_pos
+        self.assertEqual(self.device.position, max_pos)
+
+    def test_set_position_to_negative(self):
+        with self.assertRaisesRegex(Exception, "can't move to position"):
+            self.device.position = -1
+
+    def test_set_position_above_limit(self):
+        with self.assertRaisesRegex(Exception, "can't move to position"):
+            self.device.position = self.device.n_positions
 
 
 class DeformableMirrorTests(DeviceTests):
@@ -294,9 +303,10 @@ class DeformableMirrorTests(DeviceTests):
         `fake`: an object with the method `get_current_pattern`
     """
 
-    def assertCurrentPattern(self, expected_pattern, msg=''):
-        numpy.testing.assert_array_equal(self.fake.get_current_pattern(),
-                                         expected_pattern, msg)
+    def assertCurrentPattern(self, expected_pattern, msg=""):
+        numpy.testing.assert_array_equal(
+            self.fake.get_current_pattern(), expected_pattern, msg
+        )
 
     def test_get_number_of_actuators(self):
         self.assertIsInstance(self.device.n_actuators, int)
@@ -309,8 +319,8 @@ class DeformableMirrorTests(DeviceTests):
         self.assertCurrentPattern(pattern)
 
     def test_out_of_range_pattern(self):
-        ## While we expect values in the [0 1] range, we should not
-        ## actually be checking for that.
+        # While we expect values in the [0 1] range, we should not
+        # actually be checking for that.
         pattern = numpy.zeros((self.planned_n_actuators,))
         for v in [-1000, -1, 0, 1, 3]:
             pattern[:] = v
@@ -326,7 +336,7 @@ class DeformableMirrorTests(DeviceTests):
             self.assertCurrentPattern(patterns[i])
 
     def test_validate_pattern_too_long(self):
-        patterns = numpy.zeros((self.planned_n_actuators +1))
+        patterns = numpy.zeros((self.planned_n_actuators + 1))
         with self.assertRaisesRegex(Exception, "length of second dimension"):
             self.device.apply_pattern(patterns)
 
@@ -337,7 +347,9 @@ class DeformableMirrorTests(DeviceTests):
 
     def test_validate_pattern_with_extra_dimension(self):
         patterns = numpy.zeros((2, 1, self.planned_n_actuators))
-        with self.assertRaisesRegex(Exception, "dimensions \(must be 1 or 2\)"):
+        with self.assertRaisesRegex(
+            Exception, "dimensions \\(must be 1 or 2\\)"
+        ):
             self.device.apply_pattern(patterns)
 
 
@@ -349,71 +361,81 @@ class DSPTests(DeviceTests):
     pass
 
 
-class TestDummyLaser(unittest.TestCase, LaserTests):
+class TestDummyLightSource(unittest.TestCase, LightSourceTests):
     def setUp(self):
-        self.device = dummies.TestLaser()
+        self.device = dummies.TestLightSource()
 
-        ## TODO: we need to rethink the test so this is not needed.
+        # TODO: we need to rethink the test so this is not needed.
         self.fake = self.device
         self.fake.default_power = self.fake._set_point
         self.fake.min_power = 0.0
         self.fake.max_power = 100.0
 
     def test_being(self):
-        ## TODO: this test uses is_alive but that's actually a method
-        ## of SerialDeviceMixIn and not specific to lasers.  It is not
-        ## implemented on our dummy laser.  We need to decide what to
-        ## do about it.  Is this general enough that should go to all
-        ## devices?
+        # TODO: this test uses is_alive but that's actually a method
+        # of SerialDeviceMixin and not specific to lasers.  It is not
+        # implemented on our dummy laser.  We need to decide what to
+        # do about it.  Is this general enough that should go to all
+        # devices?
         pass
 
     def test_get_is_on(self):
-        ## TODO: this test assumes the connection property to be the
-        ## fake.  We need to rethink how the mock lasers work.
+        # TODO: this test assumes the connection property to be the
+        # fake.  We need to rethink how the mock lasers work.
         pass
 
 
-class TestCoherentSapphireLaser(unittest.TestCase, LaserTests,
-                                SerialDeviceTests):
+class TestCoherentSapphireLaser(
+    unittest.TestCase, LightSourceTests, SerialDeviceTests
+):
     def setUp(self):
         from microscope.lasers.sapphire import SapphireLaser
         from microscope.testsuite.mock_devices import CoherentSapphireLaserMock
-        with unittest.mock.patch('microscope.lasers.sapphire.serial.Serial',
-                                 new=CoherentSapphireLaserMock):
-            self.device = SapphireLaser('/dev/null')
+
+        with unittest.mock.patch(
+            "microscope.lasers.sapphire.serial.Serial",
+            new=CoherentSapphireLaserMock,
+        ):
+            self.device = SapphireLaser("/dev/null")
         self.device.initialize()
 
         self.fake = CoherentSapphireLaserMock
 
 
-class TestCoboltLaser(unittest.TestCase, LaserTests, SerialDeviceTests):
+class TestCoboltLaser(unittest.TestCase, LightSourceTests, SerialDeviceTests):
     def setUp(self):
         from microscope.lasers.cobolt import CoboltLaser
         from microscope.testsuite.mock_devices import CoboltLaserMock
-        with unittest.mock.patch('microscope.lasers.cobolt.serial.Serial',
-                                 new=CoboltLaserMock):
-            self.device = CoboltLaser('/dev/null')
+
+        with unittest.mock.patch(
+            "microscope.lasers.cobolt.serial.Serial", new=CoboltLaserMock
+        ):
+            self.device = CoboltLaser("/dev/null")
         self.device.initialize()
 
         self.fake = CoboltLaserMock
 
 
-class TestOmicronDeepstarLaser(unittest.TestCase, LaserTests,
-                               SerialDeviceTests):
+class TestOmicronDeepstarLaser(
+    unittest.TestCase, LightSourceTests, SerialDeviceTests
+):
     def setUp(self):
         from microscope.lasers.deepstar import DeepstarLaser
         from microscope.testsuite.mock_devices import OmicronDeepstarLaserMock
-        with unittest.mock.patch('microscope.lasers.deepstar.serial.Serial',
-                                 new=OmicronDeepstarLaserMock):
-            self.device = DeepstarLaser('/dev/null')
+
+        with unittest.mock.patch(
+            "microscope.lasers.deepstar.serial.Serial",
+            new=OmicronDeepstarLaserMock,
+        ):
+            self.device = DeepstarLaser("/dev/null")
         self.device.initialize()
 
         self.fake = OmicronDeepstarLaserMock
 
     def test_weird_initial_state(self):
-        ## The initial state of the laser may not be ideal to actual
-        ## turn it on, so test that weird settings are reset to
-        ## something adequate.
+        # The initial state of the laser may not be ideal to actual
+        # turn it on, so test that weird settings are reset to
+        # something adequate.
 
         self.device.connection.internal_peak_power = False
         self.device.connection.bias_modulation = True
@@ -436,9 +458,9 @@ class TestDummyCamera(unittest.TestCase, CameraTests):
 
 class TestImageGenerator(unittest.TestCase):
     def test_non_square_patterns_shape(self):
-        ## TODO: we should also be testing this via the camera but the
-        ## TestCamera is only square.  In the mean time, we only test
-        ## directly the _ImageGenerator.
+        # TODO: we should also be testing this via the camera but the
+        # TestCamera is only square.  In the mean time, we only test
+        # directly the _ImageGenerator.
         width = 16
         height = 32
         generator = dummies._ImageGenerator()
@@ -452,23 +474,48 @@ class TestImageGenerator(unittest.TestCase):
                 self.assertEqual(array.shape, (height, width))
 
 
-class TestEmptyDummyFilterWheel(unittest.TestCase, FilterWheelTests):
+class TestDummyController(unittest.TestCase, ControllerTests):
     def setUp(self):
-        self.device = dummies.TestFilterWheel()
+        self.laser = dummies.TestLightSource()
+        self.filterwheel = dummies.TestFilterWheel(positions=6)
+        self.device = dummies.TestController(
+            {"laser": self.laser, "filterwheel": self.filterwheel}
+        )
+
+    def test_device_names(self):
+        self.assertSetEqual(
+            {"laser", "filterwheel"}, set(self.device.devices.keys())
+        )
+
+    def test_control_filterwheel(self):
+        self.assertEqual(self.device.devices["filterwheel"].position, 0)
+        self.device.devices["filterwheel"].position = 2
+        self.assertEqual(self.device.devices["filterwheel"].position, 2)
+
+    def test_control_laser(self):
+        self.assertEqual(self.device.devices["laser"].power, 0.0)
+        self.device.devices["laser"].enable()
+        self.device.devices["laser"].power = 0.8
+        self.assertEqual(self.device.devices["laser"].power, 0.8)
 
 
-class TestOneFilterDummyFilterWheel(unittest.TestCase, FilterWheelTests):
+class TestEmptyDummyFilterWheel(unittest.TestCase):
+    def test_zero_positions(self):
+        with self.assertRaisesRegex(
+            ValueError, "positions must be a positive number"
+        ):
+            dummies.TestFilterWheel(positions=0)
+
+
+class TestOnePositionFilterWheel(unittest.TestCase, FilterWheelTests):
     def setUp(self):
-        self.device = dummies.TestFilterWheel(filters=[(0, 'DAPI', '430')])
+        self.device = dummies.TestFilterWheel(positions=1)
 
-class TestMultiFilterDummyFilterWheel(unittest.TestCase, FilterWheelTests):
-    def setUp(self):
-        self.device = dummies.TestFilterWheel(filters=[(0, 'DAPI', '430'),
-                                                       (1, 'GFP', '580'),])
 
-class TestEmptySixPositionFilterWheel(unittest.TestCase, FilterWheelTests):
+class TestSixPositionFilterWheel(unittest.TestCase, FilterWheelTests):
     def setUp(self):
         self.device = dummies.TestFilterWheel(positions=6)
+
 
 class TestDummyDeformableMirror(unittest.TestCase, DeformableMirrorTests):
     def setUp(self):
@@ -492,18 +539,20 @@ class TestBaseDevice(unittest.TestCase):
         """Unexpected kwargs on constructor raise exception.
 
         Test first that we can construct the device.  Then test that
-        it fails if there's a typo on the argument.  See issue #84.
+        it fails if there are unused kwargs.  This is an issue when
+        there are default arguments, there's a typo on the argument
+        name, and the class uses the default instead of an error.  See
+        issue #84.
         """
-        filters = [(0, 'DAPI', '430')]
-        dummies.TestFilterWheel(filters=filters)
-        ## XXX: Device.__del__ calls shutdown().  However, if __init__
-        ## failed the device is not complete and shutdown() fails
-        ## because the logger has not been created.  See comments on
-        ## issue #69.  patch __del__ to workaround this issue.
-        with unittest.mock.patch('microscope.devices.Device.__del__'):
-            with self.assertRaisesRegex(TypeError, "argument 'filteres'"):
-                dummies.TestFilterWheel(filteres=filters)
+        dummies.TestLightSource()
+        # XXX: Device.__del__ calls shutdown().  However, if __init__
+        # failed the device is not complete and shutdown() fails
+        # because the logger has not been created.  See comments on
+        # issue #69.  patch __del__ to workaround this issue.
+        with unittest.mock.patch("microscope.devices.Device.__del__"):
+            with self.assertRaisesRegex(TypeError, "argument 'power'"):
+                dummies.TestLightSource(power=2)
 
 
-if __name__ == '__main__':
+if __name__ == "__main__":
     unittest.main()
